### 2023-04-29 #5424 @editorialbot
[PRE REVIEW]: ElectricGrid.jl - A Julia-based modeling and simulation tool for power electronics-driven electric energy grids
**Submitting author:** @Webbah (<a href="http://orcid.org/0000-0003-3367-5998">Daniel Weber</a>)
**Repository:** https://github.com/upb-lea/ElectricGrid.jl
**Branch with paper.md** (empty if default branch): main
**Version:** v1.0.0
