+++ src/power_system_theory.jl
@@ -607,43 +607,44 @@ from all load and the total power provided by the sources. Thereby, steady state
 -  s_load_total::float`: Total apparent power drawn by all loads.
 - `s_source_total::float`: Total apparent power provided by all sources in steady state.
 """
-function CheckPowerBalance(parameters, num_source, num_load, CM)
+function CheckPowerBalance(parameters)
 
-    num_nodes = num_source + num_load
-    num_cables = Int(maximum(CM))
+    num_source = length(parameters["source"])
+    num_load = length(parameters["load"])
 
     p_load_total = 0
     q_load_total = 0
     s_source_total = 0
 
-    for i = 1:num_nodes
-        if i <= num_source
+    for i = 1:num_source
+
+        if parameters["source"][i]["mode"] in ["PQ", 3]
+            if parameters["source"][i]["p_set"] < 0
+                p_load_total = p_load_total - parameters["source"][i]["p_set"] / parameters["grid"]["phase"] # minus since p_set is already negative!
+            end
+
+            if parameters["source"][i]["q_set"] < 0
+                q_load_total = q_load_total - parameters["source"][i]["q_set"] / parameters["grid"]["phase"]
+            end
+
+        elseif parameters["source"][i]["mode"] in ["PV", 4]
+            if parameters["source"][i]["p_set"] < 0
+                p_load_total = p_load_total - parameters["source"][i]["p_set"] / parameters["grid"]["phase"]
+            end
             s_source_total = s_source_total + parameters["source"][i]["pwr"] / parameters["grid"]["phase"]
-
-            if parameters["source"][i]["mode"] in ["PQ", 3]
-                if parameters["source"][i]["p_set"] < 0
-                    p_load_total = p_load_total + parameters["source"][i]["p_set"] / parameters["grid"]["phase"]
-                end
-
-                if parameters["source"][i]["q_set"] < 0
-                    q_load_total = q_load_total + parameters["source"][i]["q_set"] / parameters["grid"]["phase"]
-                end
-
-            elseif parameters["source"][i]["mode"] in ["PV", 4]
-                if parameters["source"][i]["p_set"] < 0
-                    p_load_total = p_load_total + parameters["source"][i]["p_set"] / parameters["grid"]["phase"]
-                end
-            end
-
-            if haskey(parameters["source"][i], "load")
-                p_load_total += -parameters["source"][i]["pwr"] / parameters["grid"]["phase"] * 1 / 3   #TODO: V2 setdepending on setpoints
-                q_load_total += -parameters["source"][i]["pwr"] / parameters["grid"]["phase"] * 1 / 100
-            end
-
         else
-            p_load_total = p_load_total + parameters["load"][i-num_source]["pwr"] / parameters["grid"]["phase"]
-            q_load_total = q_load_total + parameters["load"][i-num_source]["pwr"] / parameters["grid"]["phase"]
+            s_source_total = s_source_total + parameters["source"][i]["pwr"] / parameters["grid"]["phase"]
         end
+
+        if haskey(parameters["source"][i], "load")
+            p_load_total += parameters["source"][i]["pwr"] / parameters["grid"]["phase"] * 1 / 3   #TODO: V2 setdepending on setpoints
+            q_load_total += parameters["source"][i]["pwr"] / parameters["grid"]["phase"] * 1 / 100
[…]
