+++ docs/JOSS/paper.md
@@ -33,7 +33,7 @@ affiliations:
   - name: Chair of Power Electronics and Electrical Drives, Paderborn University, Paderborn, Germany
     index: 1
   - name: Chair of Data Science for Engineering, Paderborn University, Paderborn, Germany
-    index: 1
+    index: 2
 date: 28 April 2023
 bibliography: paper.bib
 ---
