+++ examples/notebooks/readme.md
@@ -10,12 +10,10 @@ Nodeconstructor:
 * [The theoretical principles behind the calculations](https://github.com/upb-lea/ElectricGrid.jl/blob/main/examples/notebooks/NodeConstructor_Theory_DEMO.ipynb)
 * [Background of automated energy grid configuration](https://github.com/upb-lea/ElectricGrid.jl/blob/main/examples/notebooks/NodeConstructor_Application_DEMO.ipynb)
 
-* [Overview of parameters usable for experiment configuration](https://github.com/upb-lea/ElectricGrid.jl/blob/main/examples/notebooks/Default_Parameters.ipynb)
-
 Classic controllers:
 * [Auxiliaries provided from the classic control side](https://github.com/upb-lea/ElectricGrid.jl/blob/main/examples/notebooks/1_Auxiliaries_OU_process.ipynb)
 
-* [Open-loop swing mode - infinite bus](https://github.com/upb-lea/ElectricGrid.jl/blob/main/examples/notebooks/RL_Complex_DEMO.ipynb)
+* [Open-loop swing mode - infinite bus](https://github.com/upb-lea/ElectricGrid.jl/blob/develop/examples/notebooks/1_Classical_Controllers_Swing.ipynb)
 
 * [PQ mode - controllable source and load](https://github.com/upb-lea/ElectricGrid.jl/blob/main/examples/notebooks/2_Classical_Controllers_PQ.ipynb)
 
@@ -24,7 +22,7 @@ Classic controllers:
 * [VSG mode - enhanced grid forming](https://github.com/upb-lea/ElectricGrid.jl/blob/main/examples/notebooks/4_Classical_Controllers_VSG.ipynb)
 
 Reinforcement Learning:
-* [Train an RL agent in the ElectricGrid.jl framework](https://github.com/upb-lea/ElectricGrid.jl/blob/main/examples/notebooks/NodeConstructor_Theory_DEMO.ipynb)
+* [Train an RL agent in the ElectricGrid.jl framework](https://github.com/upb-lea/ElectricGrid.jl/blob/develop/examples/notebooks/RL_Single_Agent_DEMO.ipynb)
 
 * [Train an RL agent in a classicly controlled grid](https://github.com/upb-lea/ElectricGrid.jl/blob/main/examples/notebooks/RL_Classical_Controllers_Merge_DEMO.ipynb)
 
