### 2023-05-05 5d3c83d
finished septimus code section in data_hook_2
