+++ src/data_hook_2.jl
@@ -218,7 +218,6 @@ function (hook::DataHook2)(::PreActStage, agent, env, action, training = false)
         hook.policy_names = [s for s in keys(agent.agents)]
     end
 
-    #TODO: append_tmp for Septimus Code
     if findfirst(x -> x == "classic", hook.policy_names) !== nothing
 
         ClassicalPolicy = agent.agents["classic"]["policy"].policy
@@ -235,43 +234,60 @@ function (hook::DataHook2)(::PreActStage, agent, env, action, training = false)
         for idx in hook.v_dq
             s_idx = findfirst(x -> x == idx, ClassicalPolicy.Source_Indices)
             if s_idx !== nothing
-                hook.list_iterator == 0 && push!(hook.column_names, Symbol("source$(idx)_v_d"))
-                push!(hook.tmp, ClassicalPolicy.Source.V_dq0[s_idx, 1])
-
-                hook.list_iterator == 0 && push!(hook.column_names, Symbol("source$(idx)_v_q"))
-                push!(hook.tmp, ClassicalPolicy.Source.V_dq0[s_idx, 2])
+                if hook.list_iterator == 0
+                    push!(hook.column_names, Symbol("source$(idx)_v_d"))
+                    push!(hook.tmp, ClassicalPolicy.Source.V_dq0[s_idx, 1])
+                    push!(hook.column_names, Symbol("source$(idx)_v_q"))
+                    push!(hook.tmp, ClassicalPolicy.Source.V_dq0[s_idx, 2])
+                else
+                    append_tmp(hook, ClassicalPolicy.Source.V_dq0[s_idx, 1])
+                    append_tmp(hook, ClassicalPolicy.Source.V_dq0[s_idx, 2])
+                end
             end
         end
 
         for idx in hook.i_dq
             s_idx = findfirst(x -> x == idx, ClassicalPolicy.Source_Indices)
             if s_idx !== nothing
-                hook.list_iterator == 0 && push!(hook.column_names, Symbol("source$(idx)_i_d"))
-                push!(hook.tmp, ClassicalPolicy.Source.I_dq0[s_idx, 1])
-
-                hook.list_iterator == 0 && push!(hook.column_names, Symbol("source$(idx)_i_q"))
-                push!(hook.tmp, ClassicalPolicy.Source.I_dq0[s_idx, 2])
+                if hook.list_iterator == 0
+                    push!(hook.column_names, Symbol("source$(idx)_i_d"))
+                    push!(hook.tmp, ClassicalPolicy.Source.I_dq0[s_idx, 1])
+                    push!(hook.column_names, Symbol("source$(idx)_i_q"))
+                    push!(hook.tmp, ClassicalPolicy.Source.I_dq0[s_idx, 2])
+                else
+                    append_tmp(hook, ClassicalPolicy.Source.I_dq0[s_idx, 1])
+                    append_tmp(hook, ClassicalPolicy.Source.I_dq0[s_idx, 2])
+                end
             end
         end
 
         for idx in hook.power_pq_inv
             s_idx = findfirst(x -> x == idx, ClassicalPolicy.Source_Indices)
             if s_idx !== nothing
-                hook.list_iterator == 0 && push!(hook.column_names, Symbol("source$(idx)_p_inv"))
-                push!(hook.tmp, ClassicalPolicy.Source.p_q_inv[s_idx, 1])
-
-                hook.list_iterator == 0 && push!(hook.column_names, Symbol("source$(idx)_q_inv"))
-                push!(hook.tmp, ClassicalPolicy.Source.p_q_inv[s_idx, 2])
+                if hook.list_iterator == 0
+                    push!(hook.column_names, Symbol("source$(idx)_p_inv"))
+                    push!(hook.tmp, ClassicalPolicy.Source.p_q_inv[s_idx, 1])
[…]
