+++ src/classical_control_2.jl
@@ -882,49 +882,62 @@ calculating the actions.
 function ClassicalControl(Animo::ClassicalPolicy, env::ElectricGridEnv)
 
     Source = Animo.Source
-    SourceInterface(Animo, env)
+
+    @timeit to "SourceInterface" begin
+        SourceInterface(Animo, env)
+    end
 
     ramp_end = Source.ramp_end
 
-    OrnsteinUhlenbeck(Source)
+    @timeit to "OrnsteinUhlenbeck" begin
+        OrnsteinUhlenbeck(Source)
+    end
 
-    Threads.@threads for ns in 1:Source.num_sources
+    @timeit to "Loop Sources" begin
+    #Threads.@threads
+        for ns in 1:Source.num_sources
 
-        if Source.Source_Modes[ns] == "Swing"
+            if Source.Source_Modes[ns] == "Swing"
 
-            SwingMode(Source, ns, t_end = ramp_end)
-        elseif Source.Source_Modes[ns] == "Voltage"
+                SwingMode(Source, ns, t_end = ramp_end)
+            elseif Source.Source_Modes[ns] == "Voltage"
 
-            VoltageControlMode(Source, ns, t_end = ramp_end)
-        elseif Source.Source_Modes[ns] == "PQ"
+                VoltageControlMode(Source, ns, t_end = ramp_end)
+            elseif Source.Source_Modes[ns] == "PQ"
 
-            PQControlMode(Source, ns, Source.pq0_set[ns, :])
-        elseif Source.Source_Modes[ns] == "PV"
+                PQControlMode(Source, ns, Source.pq0_set[ns, :])
+            elseif Source.Source_Modes[ns] == "PV"
 
-            PVControlMode(Source, ns, Source.pq0_set[ns, :])
-        elseif Source.Source_Modes[ns] == "Droop" || Source.Source_Modes[ns] == "Semi-Droop"
+                PVControlMode(Source, ns, Source.pq0_set[ns, :])
+            elseif Source.Source_Modes[ns] == "Droop" || Source.Source_Modes[ns] == "Semi-Droop"
 
-            DroopControlMode(Source, ns, t_end = ramp_end)
-        elseif Source.Source_Modes[ns] == "Synchronverter"
+                DroopControlMode(Source, ns, t_end = ramp_end)
+            elseif Source.Source_Modes[ns] == "Synchronverter"
 
-            SynchronverterMode(Source, ns, pq0_ref = Source.pq0_set[ns, :], mode = 1, t_end = ramp_end)
-        elseif Source.Source_Modes[ns] == "Semi-Synchronverter"
+                SynchronverterMode(Source, ns, pq0_ref = Source.pq0_set[ns, :], mode = 1, t_end = ramp_end)
+            elseif Source.Source_Modes[ns] == "Semi-Synchronverter"
 
-            SynchronverterMode(Source, ns, pq0_ref = Source.pq0_set[ns, :], mode = 2, t_end = ramp_end)
-        elseif Source.Source_Modes[ns] == "Step"
+                SynchronverterMode(Source, ns, pq0_ref = Source.pq0_set[ns, :], mode = 2, t_end = ramp_end)
+            elseif Source.Source_Modes[ns] == "Step"
 
-            StepMode(Source, ns, t_end = ramp_end)
-        elseif Source.Source_Modes[ns] == "Not Used 2"
+                StepMode(Source, ns, t_end = ramp_end)
+            elseif Source.Source_Modes[ns] == "Not Used 2"
 
[…]
