+++ src/classical_control_2.jl
@@ -964,12 +964,10 @@ function SourceInterface(Animo::ClassicalPolicy, env::ElectricGridEnv)
 
     Source = Animo.Source
     Source.steps = env.steps + 1
-    ω = 2π*Source.fsys
-    Source.θsys = (Source.θsys + Source.ts*ω)%(2π)
+    Source.θsys = (Source.θsys + Source.ts*2π*Source.fsys)%(2π)
 
     state = env.x[findall(x -> x in Animo.state_ids, env.state_ids)]
 
-    #Threads.@threads
     for ns in 1:Source.num_sources
 
         #= # for single phase measurements
@@ -979,25 +977,21 @@ function SourceInterface(Animo::ClassicalPolicy, env::ElectricGridEnv)
         =#
 
         #----------------------------------------------------------------------------------
-
-        @timeit to "Rolling Windows" begin
-        # Rolling Windows
-            roll(Source.Vd_abc_new, ns)
-            roll(Source.V_filt_poc, ns)
-            roll(Source.I_filt_inv, ns)
-
-        end
-
-        #----------------------------------------------------------------------------------
         # New values from environment
         @timeit to "New values" begin
-            update_value(Source.I_filt_inv, ns, state[Source.I_inv_loc[:, ns]])
 
             if Source.filter_type[ns] == "LCL"
 
                 if Source.Observer[ns]
-                    update_value(Source.V_filt_poc, ns, state[Source.V_cable_loc[:, ns]])
+
+                    roll(Source.Vd_abc_new, ns)
+                    roll(Source.I_filt_inv, ns)
+
+                    update_value(Source.I_filt_inv, ns, state[Source.I_inv_loc[:, ns]])
+
                     LuenbergerObserver(Source, ns)
+                    update_value(Source.V_filt_poc, ns, state[Source.V_cable_loc[:, ns]])
+
                 else
 
                     update_value(Source.V_filt_poc, ns, state[Source.V_cable_loc[:, ns]])
@@ -1006,10 +1000,14 @@ function SourceInterface(Animo::ClassicalPolicy, env::ElectricGridEnv)
                     update_value(Source.V_filt_cap, ns, state[Source.V_cap_loc[:, ns]])
                 end
 
+                update_value(Source.I_filt_inv, ns, state[Source.I_inv_loc[:, ns]])
+
                 @views icap = Source.I_filt_inv[ns, :, end] .- state[Source.I_poc_loc[:, ns]]
                 update_value(Source.V_filt_cap, ns, Source.V_filt_cap[ns, :, end] .+ Source.Rf_C[ns]*icap)
 
             elseif Source.filter_type[ns] == "LC"
+
+                update_value(Source.I_filt_inv, ns, state[Source.I_inv_loc[:, ns]])
 
                 update_value(Source.V_filt_poc, ns, Source.V_filt_cap[ns, :, end])
                 update_value(Source.I_filt_poc, ns, Source.I_filt_inv[ns, :, end] .- env.y[Source.I_poc_loc[:, ns]])
@@ -1019,22 +1017,25 @@ function SourceInterface(Animo::ClassicalPolicy, env::ElectricGridEnv)
 
             elseif Source.filter_type[ns] == "L"
 
+                update_value(Source.I_filt_inv, ns, state[Source.I_inv_loc[:, ns]])
+
                 update_value(Source.V_filt_cap, ns, state[Source.V_cable_loc[:, ns]])
 
[…]
