+++ performance_test/test_classical_2.jl
@@ -20,7 +20,7 @@ end
 parameters = Dict{Any, Any}(
         "source" => Any[
                         Dict{Any, Any}("pwr" => 400e3,
-                                        "mode" => "VSG",
+                                        "mode" => "Swing",
                                         "fltr" => "LC",
                                         "v_δ_set" => 4),
                         Dict{Any, Any}("pwr" => 100e3,
