+++ src/classical_control_2.jl
@@ -44,6 +44,8 @@ mutable struct ClassicalControls
     Pm::Matrix{Float64}
     Qm::Matrix{Float64}
 
+    power_mat::Matrix{Float64}
+
     debug::Vector{Float64}
 
     #---------------------------------------------------------------------------
@@ -216,7 +218,7 @@ mutable struct ClassicalControls
         Lf_1::Vector{Float64}, Lf_2::Vector{Float64}, Cf::Vector{Float64},
         Rf_L1::Vector{Float64}, Rf_L2::Vector{Float64}, Rf_C::Vector{Float64},
         T_eval::Int64, T_sp_rms::Float64, V_ph::Array{Float64}, I_ph::Array{Float64},
-        p_q_inst::Matrix{Float64}, p_inst::Matrix{Float64}, Pm::Matrix{Float64}, Qm::Matrix{Float64},
+        p_q_inst::Matrix{Float64}, p_inst::Matrix{Float64}, Pm::Matrix{Float64}, Qm::Matrix{Float64}, power_mat::Matrix{Float64},
         debug::Vector{Float64}, Modes::Dict{String, Int64}, Source_Modes::Vector{String},
         num_sources::Int64, phases::Int64,
         f_cntr::Float64, fsys::Float64, θsys::Float64,
@@ -263,7 +265,7 @@ mutable struct ClassicalControls
         Lf_1, Lf_2, Cf,
         Rf_L1, Rf_L2, Rf_C,
         T_eval, T_sp_rms, V_ph, I_ph,
-        p_q_inst, p_inst, Pm, Qm,
+        p_q_inst, p_inst, Pm, Qm, power_mat,
         debug, Modes, Source_Modes,
         num_sources, phases,
         f_cntr, fsys, θsys,
@@ -376,6 +378,9 @@ mutable struct ClassicalControls
         Pm = fill!(Pm, 0)
         Qm = Array{Float64, 2}(undef, num_sources, phases+1) # 4th column is total
         Qm = fill!(Qm, 0)
+
+        power_mat = Array{Float64, 2}(undef, 3, 3)
+        power_mat = fill!(power_mat, 0.0)
 
         debug = Array{Float64, 1}(undef, 20) # put anything in here that needs to be debugged - for plotting
         debug = fill!(debug, 0)
@@ -636,7 +641,7 @@ mutable struct ClassicalControls
         Lf_1, Lf_2, Cf,
         Rf_L1, Rf_L2, Rf_C,
         T_eval, T_sp_rms, V_ph, I_ph,
-        p_q_inst, p_inst, Pm, Qm,
+        p_q_inst, p_inst, Pm, Qm, power_mat,
         debug, Modes, Source_Modes,
         num_sources, phases,
         f_cntr, fsys, θsys,
@@ -976,15 +981,17 @@ function SourceInterface(Animo::ClassicalPolicy, env::ElectricGridEnv)
         Source.I_filt_poc[ns, :, 1:end-1] = Source.I_filt_poc[ns, :, 2:end]
         =#
 
+        roll(Source.Vd_abc_new, ns)
+
         #----------------------------------------------------------------------------------
         # New values from environment
-        @timeit to "New values" begin
+        #@timeit to "New values" begin
 
             if Source.filter_type[ns] == "LCL"
 
                 if Source.Observer[ns]
 
-                    roll(Source.Vd_abc_new, ns)
+                    #roll(Source.Vd_abc_new, ns)
                     roll(Source.I_filt_inv, ns)
 
[…]
