### 2023-05-08 ddac3a4
add missing zero_expression
