+++ src/ElectricGrid.jl
@@ -36,7 +36,7 @@ include("./custom_control.jl")
 include("./solar_module.jl")
 include("./electric_grid_env.jl")
 include("./agent_ddpg.jl")
-include("./classical_control_2.jl")
+include("./classical_control_3.jl")
 include("./multi_controller.jl")
 include("./render.jl")
 include("./data_hook.jl")
