+++ src/classical_control_3.jl
@@ -887,6 +887,10 @@ calculating the actions.
 function ClassicalControl(Animo::ClassicalPolicy, env::ElectricGridEnv)
 
     Source = Animo.Source
+    Source.steps = env.steps + 1
+    Source.θsys = (Source.θsys + Source.ts*2π*Source.fsys)%(2π)
+
+    state = env.x[findall(x -> x in Animo.state_ids, env.state_ids)]
 
     ramp_end = Source.ramp_end
 
@@ -894,85 +898,7 @@ function ClassicalControl(Animo::ClassicalPolicy, env::ElectricGridEnv)
         OrnsteinUhlenbeck(Source)
     end
 
-    @timeit to "SourceInterface" begin
-        SourceInterface(Animo, env)
-    end
-
-    @timeit to "Loop Sources" begin
     #Threads.@threads
-        for ns in 1:Source.num_sources
-
-            if Source.Source_Modes[ns] == "Swing"
-
-                SwingMode(Source, ns, t_end = ramp_end)
-            elseif Source.Source_Modes[ns] == "Voltage"
-
-                VoltageControlMode(Source, ns, t_end = ramp_end)
-            elseif Source.Source_Modes[ns] == "PQ"
-
-                PQControlMode(Source, ns, Source.pq0_set[ns, :])
-            elseif Source.Source_Modes[ns] == "PV"
-
-                PVControlMode(Source, ns, Source.pq0_set[ns, :])
-            elseif Source.Source_Modes[ns] == "Droop" || Source.Source_Modes[ns] == "Semi-Droop"
-
-                DroopControlMode(Source, ns, t_end = ramp_end)
-            elseif Source.Source_Modes[ns] == "Synchronverter"
-
-                SynchronverterMode(Source, ns, pq0_ref = Source.pq0_set[ns, :], mode = 1, t_end = ramp_end)
-            elseif Source.Source_Modes[ns] == "Semi-Synchronverter"
-
-                SynchronverterMode(Source, ns, pq0_ref = Source.pq0_set[ns, :], mode = 2, t_end = ramp_end)
-            elseif Source.Source_Modes[ns] == "Step"
-
-                StepMode(Source, ns, t_end = ramp_end)
-            elseif Source.Source_Modes[ns] == "Not Used 2"
-
-                SynchronverterMode(Source, ns, pq0_ref = Source.pq0_set[ns, :], mode = 2, t_end = ramp_end)
-            elseif Source.Source_Modes[ns] == "Not Used 3"
-
-                SynchronverterMode(Source, ns, pq0_ref = Source.pq0_set[ns, :], mode = 2, t_end = ramp_end)
-            end
-        end
-    end
-
-    @timeit to "EnvInterface" begin
-        Action = EnvInterface(Source)
-    end
-
-    @timeit to "Measurements" begin
-        Measurements(Source)
-    end
-
-    return Action
-end
-
-function roll(matrix, ns = nothing)
-    if isnothing(ns)
-        @views matrix[:, 1:end-1] = matrix[:, 2:end]
-    else
-        @views matrix[ns, :, 1:end-1] = matrix[ns, :, 2:end]
-    end
-end
-
-function update_value(matrix, ns, vector)
-    @views matrix[ns, :, end] = vector
-end
-
-"""
-    SourceInterface(ClassicalPolicy, ElectricGridEnv)
-
-# Description
-"Measures" or "Observes" the relevant quantities necessary for control.
-"""
-function SourceInterface(Animo::ClassicalPolicy, env::ElectricGridEnv)
-
-    Source = Animo.Source
-    Source.steps = env.steps + 1
[…]
