+++ src/classical_control_3.jl
@@ -917,49 +917,51 @@ function ClassicalControl(Animo::ClassicalPolicy, env::ElectricGridEnv)
 
                 if Source.Observer[ns]
 
-                    roll(Source.I_filt_inv, ns)
+                    I_filt_poc_new, V_filt_cap_new = LuenbergerObserver(Source, ns, state[Source.I_inv_loc[:, ns]], Source.I_filt_inv[ns, :, end], Source.V_filt_poc[ns, :, end])
 
-                    update_value(Source.I_filt_inv, ns, state[Source.I_inv_loc[:, ns]])
+                    V_filt_cap_new = V_filt_cap_new .+ Source.Rf_C[ns]*(state[Source.I_inv_loc[:, ns]] .- state[Source.I_poc_loc[:, ns]])
+                    V_filt_poc_new = state[Source.V_cable_loc[:, ns]]
+                    I_filt_inv_new = state[Source.I_inv_loc[:, ns]]
 
-                    LuenbergerObserver(Source, ns)
-                    update_value(Source.V_filt_poc, ns, state[Source.V_cable_loc[:, ns]])
+                    update_value(Source.V_filt_poc, ns, V_filt_poc_new)
+                    update_value(Source.I_filt_inv, ns, I_filt_inv_new)
 
                 else
 
-                    update_value(Source.V_filt_poc, ns, state[Source.V_cable_loc[:, ns]])
+                    V_filt_cap_new = state[Source.V_cap_loc[:, ns]] .+ Source.Rf_C[ns]*(state[Source.I_inv_loc[:, ns]] .- state[Source.I_poc_loc[:, ns]])
+                    I_filt_poc_new = state[Source.I_poc_loc[:, ns]]
+                    V_filt_poc_new = state[Source.V_cable_loc[:, ns]]
+                    I_filt_inv_new = state[Source.I_inv_loc[:, ns]]
 
-                    update_value(Source.I_filt_poc, ns, state[Source.I_poc_loc[:, ns]])
-                    update_value(Source.V_filt_cap, ns, state[Source.V_cap_loc[:, ns]])
                 end
-
-                #update_value(Source.I_filt_inv, ns, state[Source.I_inv_loc[:, ns]])
-                @views Source.I_filt_inv[ns, :, end] = state[Source.I_inv_loc[:, ns]]
-
-                #@views icap = Source.I_filt_inv[ns, :, end] .- state[Source.I_poc_loc[:, ns]]
-                #update_value(Source.V_filt_cap, ns, Source.V_filt_cap[ns, :, end] .+ Source.Rf_C[ns]*icap)
-
-                @views Source.V_filt_cap[ns, :, end] = Source.V_filt_cap[ns, :, end] .+ Source.Rf_C[ns]*(Source.I_filt_inv[ns, :, end] .- state[Source.I_poc_loc[:, ns]])
 
             elseif Source.filter_type[ns] == "LC"
 
-                update_value(Source.I_filt_inv, ns, state[Source.I_inv_loc[:, ns]])
-
-                update_value(Source.V_filt_poc, ns, Source.V_filt_cap[ns, :, end])
-                update_value(Source.I_filt_poc, ns, Source.I_filt_inv[ns, :, end] .- env.y[Source.I_poc_loc[:, ns]])
-
-                @views icap = env.y[Source.I_poc_loc[:, ns]]
-                update_value(Source.V_filt_cap, ns, state[Source.V_cable_loc[:, ns]] .+ Source.Rf_C[ns]*icap)
+                V_filt_cap_new = state[Source.V_cable_loc[:, ns]] .+ Source.Rf_C[ns]*env.y[Source.I_poc_loc[:, ns]]
[…]
