### 2023-05-12 2532adc
tried the old setup and it learns
