+++ examples/Investigations/RL_Classical_Controllers_Merge_DEMO.jl
@@ -20,26 +20,12 @@ Dict{Any, Any}(
                     Dict{Any, Any}(
                         "pwr" => 200e3,
                         "control_type" => "RL",
-                        "fltr" => "LCL",),
+                        "fltr" => "L",),
                     Dict{Any, Any}(
                         "pwr" => 200e3,
-                        "fltr" => "LCL",
+                        "fltr" => "LC",
                         "control_type" => "classic",
                         "mode" => "Swing",),
-                    ],
-    "load"   => Any[
-                    Dict{Any, Any}("impedance" => "RL",
-                                    "R" => R_load,
-                                    "L" => L_load,
-                                    "v_limit" => 10e3),
-                                    ],
-    "cable"   => Any[
-                    Dict{Any, Any}("R" => 0.1,
-                                    "L" => 0.0025e-3,
-                                    "C" => 0.001e-4),
-                    Dict{Any, Any}("R" => 0.1,
-                                    "L" => 0.0025e-3,
-                                    "C" => 0.001e-4),
                     ],
     "grid" => Dict{Any, Any}(
         "phase" => 3,
@@ -50,9 +36,10 @@ Dict{Any, Any}(
 
 function reward_function(env, name = nothing)
 
-    if name == "my_ddpg"
-
-        p_ref = 0e3
+    if name == "classic"
+        return 0
+    else
+        p_ref = 1e3
         q_ref = 0e3
 
         pq0_ref = [p_ref; q_ref; 0]
@@ -60,9 +47,9 @@ function reward_function(env, name = nothing)
         v_a_meas = env.x[findfirst(x -> x == "source1_v_C_cables_a", env.state_ids)]
         v_b_meas = env.x[findfirst(x -> x == "source1_v_C_cables_b", env.state_ids)]
         v_c_meas  = env.x[findfirst(x -> x == "source1_v_C_cables_c", env.state_ids)]
-        i_a_meas = env.state[findfirst(x -> x == "source1_i_L2_a", env.state_ids)]
-        i_b_meas = env.state[findfirst(x -> x == "source1_i_L2_b", env.state_ids)]
-        i_c_meas  = env.state[findfirst(x -> x == "source1_i_L2_c", env.state_ids)]
+        i_a_meas = env.state[findfirst(x -> x == "source1_i_L1_a", env.state_ids)]
+        i_b_meas = env.state[findfirst(x -> x == "source1_i_L1_b", env.state_ids)]
+        i_c_meas  = env.state[findfirst(x -> x == "source1_i_L1_c", env.state_ids)]
 
         v_abc_meas = [v_a_meas; v_b_meas; v_c_meas]
         i_abc_meas = [i_a_meas; i_b_meas; i_c_meas]
@@ -71,38 +58,75 @@ function reward_function(env, name = nothing)
 
         i_abc_ref = clamp.(i_abc_ref/env.nc.parameters["source"][1]["i_limit"], -0.95, 0.95)
 
-        i_abc_ref= [0; 0; 0]
+        #i_abc_ref= [0; 0; 0]
 
         #println(i_abc_ref)
 
-        if any((abs.(env.state)).>=1)
-            return -1
-        end
+        # if any((abs.(env.state)).>=1)
+        #     return -1
+        # end
 
         r = 1 - 1/3*(sum((abs.(i_abc_ref - i_abc_meas)/2).^0.5))
 
-        return r * (1 - agent.policy.γ)/2
+        r += (0.2 - clamp(maximum(abs.(env.state)), 0.2, 1.0))
 
-    else
+        return r * (1 - 0.99)/2
+    end
+
+end
+
+function reference(t)
+    θ = 2*pi*50*t
+    θph = [θ; θ - 120π/180; θ + 120π/180]
[…]
