### 2023-05-12 c70feff
Droop Experiment (learning well)
