### 2023-05-16 3c58916
added Nodconstructor test for nonlinear things
