+++ Project.toml
@@ -45,39 +45,6 @@ UnicodePlots = "b8865327-cd53-5732-bb35-84acbb429228"
 Zygote = "e88e6eb3-aa80-5325-afca-941959d7151f"
 
 [compat]
-CDDLib = "0.9.2"
-CSV = "0.10.9"
-CUDA = "3.13.1"
-Combinatorics = "1.0.2"
-ControlSystemsBase = "1.4.1"
-DataFrames = "1.3.6"
-DataStructures = "0.18.13"
-Distributions = "0.25.87"
-Documenter = "0.27.24"
-Flux = "0.12.10"
-Glob = "1.3.1"
-GraphPlot = "0.5.2"
-Graphs = "1.8.0"
-IntervalSets = "0.5.4"
-Ipopt = "1.2.1"
-JSON = "0.21.4"
-JuMP = "1.10.0"
-LoggingExtras = "1.0.0"
-MAT = "0.10.4"
-MacroTools = "0.5.10"
-NLSolvers = "0.4.0"
-NonlinearSolve = "0.3.22"
-PlotlyBase = "0.8.19"
-PlotlyJS = "0.18.10"
 ReinforcementLearning = "0.10.2"
 ReinforcementLearningCore = "0.8.13"
-SciMLBase = "1.85.0"
-Setfield = "0.8.2"
-SpecialFunctions = "2.2.0"
-StableRNGs = "1.0.0"
-StaticArrays = "1.5.20"
-StatsBase = "0.33.21"
-TimerOutputs = "0.5.22"
-UnicodePlots = "2.12.4"
-Zygote = "0.6.60"
-julia = "^1.8"
+
