+++ examples/Investigations/RL_Classical_Controllers_Merge_DEMO.jl
@@ -12,7 +12,7 @@ CM = [ 0. 0. 1.
         0. 0. 2.
         -1. -2. 0.]
 
-R_load, L_load, _, _ = ParallelLoadImpedance(50e3, 0.99, 230)
+R_load, L_load, _, _ = ParallelLoadImpedance(50e3, 0.95, 230)
 
 parameters =
 Dict{Any, Any}(
@@ -20,7 +20,8 @@ Dict{Any, Any}(
                     Dict{Any, Any}(
                         "pwr" => 200e3,
                         "control_type" => "RL",
-                        "fltr" => "L",),
+                        "fltr" => "L",
+                        "i_rip" => 0.25,),
                     Dict{Any, Any}(
                         "pwr" => 200e3,
                         "fltr" => "LC",
@@ -82,7 +83,31 @@ function reference(t)
 end
 
 featurize_ddpg = function(state, env, name)
-    if name != "classic"
+    if name == "ElectricGrid_ddpg_1"
+
+        #state = state[findall(x -> split(x, "_")[2] == "i" , env.agent_dict["ElectricGrid_ddpg_1"]["state_ids"])]
+
+        norm_ref = env.nc.parameters["source"][1]["i_limit"]
+        state = vcat(state, reference(env.t)/norm_ref)
+
+        θ = 2*pi*50*env.t
+
+        state_to_control_1 = env.state[findfirst(x -> x == "source1_i_L1_a", env.state_ids)]
+        state_to_control_2 = env.state[findfirst(x -> x == "source1_i_L1_b", env.state_ids)]
+        state_to_control_3 = env.state[findfirst(x -> x == "source1_i_L1_c", env.state_ids)]
+
+        state_to_control = [state_to_control_1, state_to_control_2, state_to_control_3]
+
+        Il_dq0 = DQ0Transform(state_to_control, θ)
+        state = vcat(state, Il_dq0)
+    end
+end
+
+featurize_ddpg2 = function(state, env, name)
+    if name == "ElectricGrid_ddpg_1"
+
+        state = state[findall(x -> split(x, "_")[2] == "i" , env.agent_dict["ElectricGrid_ddpg_1"]["state_ids"])]
+
         norm_ref = env.nc.parameters["source"][1]["i_limit"]
         state = vcat(state, reference(env.t)/norm_ref)
     end
@@ -124,11 +149,147 @@ env = ElectricGridEnv(
 
 controllers = SetupAgents(env)
 
+using FileIO, JLD2
+#FileIO.save("controllers.jld2","controllers",controllers)
+#controllers = FileIO.load("controllers.jld2","controllers");
+
+learnhook = DataHook()
+
 # Let it learn till 19_000 !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
 function learn()
+    num_steps = 50_000
     while true
-        Learn(controllers, env, num_episodes = 50)
+        Learn(controllers, env, steps = num_steps, hook = learnhook)
     end
+end
+
+function learn5()
+    num_steps = 3_000_000
+
+    global rewardresults1 = [[], [], [], [], []]
+    rewardresults1 = convert(Vector{Vector{Float64}}, rewardresults1)
+    for i in 1:5
+        learnmore = true
+
+        env = ElectricGridEnv(
+            #CM =  CM,
+            parameters = parameters,
+            t_end = 1,
+            reward_function = reward_function2,
+            featurize = featurize_ddpg,
+            action_delay = 0,
+            verbosity = 0)
+
+        controllers = SetupAgents(env)
+        learnhook = DataHook()
+
+        Learn(controllers, env, steps = num_steps, hook = learnhook)
+
+        println(length(learnhook.df[!,"reward"]))
+
[…]
