+++ src/render.jl
@@ -226,8 +226,13 @@ function RenderHookResults(; hook, states_to_plot = nothing, actions_to_plot = n
             xanchor="right",
             orientation="h"
         ),
-        width = 800,
-        height = 550,
+        font=attr(
+            family="Arial",
+            size=16,
+            color="black"
+        ),
+        #width = 800,
+        #height = 550,
         margin=attr(l=100, r=80, b=80, t=100, pad=10)
     )
 
