### 2023-06-15 a8ad4ca
NodeConstructor works with functions
