+++ src/node_constructor.jl
@@ -1,18 +1,19 @@
-import Base:/,*,+,-,^
+import Base: /, *, +, -, ^
 
-/(a::Number,b::Function) = x->a/b(x)
-/(a::Function,b::Function) = x->a(x)/b(x)
-*(a::Function,b::Number) = x->a(x)*b
-*(b::Number,a::Function) = x->a(x)*b
-*(a::Function,b::Function) = x->a(x)*b(x)
-+(a::Function,b::Number) = x->a(x)+b
-+(b::Number,a::Function) = x->a(x)+b
-+(a::Function,b::Function) = x->a(x)+b(x)
--(a::Function,b::Number) = x->a(x)-b
--(b::Number,a::Function) = x->b-a(x)
--(a::Function,b::Function) = x->a(x)-b(x)
--(a::Function) = x->-a(x)
-^(a::Function,b::Number) = x->a(x)^b
+/(a::Number, b::Function) = x -> a / b(x)
+/(a::Function, b::Function) = x -> a(x) / b(x)
+/(a::Function, b::Number) = x -> a(x) / b
+*(a::Function, b::Number) = x -> a(x) * b
+*(a::Number, b::Function) = x -> a * b(x)
+*(a::Function, b::Function) = x -> a(x) * b(x)
++(a::Function, b::Number) = x -> a(x) + b
++(b::Number, a::Function) = x -> a(x) + b
++(a::Function, b::Function) = x -> a(x) + b(x)
+-(a::Function, b::Number) = x -> a(x) - b
+-(b::Number, a::Function) = x -> b - a(x)
+-(a::Function, b::Function) = x -> a(x) - b(x)
+-(a::Function) = x -> -a(x)
+^(a::Function, b::Number) = x -> a(x)^b
 
 mutable struct NodeConstructor
     num_connections
@@ -92,7 +93,7 @@ function NodeConstructor(; num_sources, num_loads, CM=nothing, parameters=nothin
             but got $(size(CM)[1]) columns and $(size(CM)[2]) rows.")
 
         @assert(-transpose(CM .* LowerTriangular(ones(size(CM)))) ==
-            CM .* UpperTriangular(ones(size(CM))),
+                CM .* UpperTriangular(ones(size(CM))),
             "Expect the CM matrix to be a antisymetric: transpose(CM) = -CM.")
 
         num_connections = Int(maximum(CM))
@@ -128,8 +129,8 @@ function NodeConstructor(; num_sources, num_loads, CM=nothing, parameters=nothin
             but got $num_loads and $(length(parameters["load"]))")
 
         @assert(length(parameters["cable"]) == num_connections,
-        "Expect the number of cables to match the number of cables in the parameters, but
-        got $num_connections and $(length(parameters["cable"]))")
+            "Expect the number of cables to match the number of cables in the parameters, but
+            got $num_connections and $(length(parameters["cable"]))")
 
         num_fltr_LCL, num_fltr_LC, num_fltr_L = CountFilters(parameters["source"])
         loads = CountLoads(parameters["load"])
@@ -147,7 +148,7 @@ function NodeConstructor(; num_sources, num_loads, CM=nothing, parameters=nothin
             $(num_fltr_LCL + num_fltr_LC + num_fltr_L)")
 
         @assert((num_loads_RLC + num_loads_LC + num_loads_RL + num_loads_RC + num_loads_R +
-            num_loads_C + num_loads_L == num_loads),
+                 num_loads_C + num_loads_L == num_loads),
             "Expect the number of loads to be identical to the sum of the loads types, but
             the number of loads is $num_loads and the sum of the loads is
[…]
