### 2023-06-15 e4c5fe7
example script for the controllers paper
