+++ examples/Investigations/RL_Classical_Controllers_Merge_DEMO.jl
@@ -1,350 +1,0 @@
-using ElectricGrid
-#using ReinforcementLearning
-
-"""
-This scipt contains the content of the RL_Classical_Controllers_Merge_DEMO.ipynb notebook.
-For comments and more documentation see refering notebookin examples/notebooks
-"""
-
-println("...........o0o----ooo0§0ooo~~~  START  ~~~ooo0§0ooo----o0o...........\n\n")
-
-CM = [ 0. 0. 1.
-        0. 0. 2.
-        -1. -2. 0.]
-
-R_load, L_load, _, _ = ParallelLoadImpedance(50e3, 0.95, 230)
-
-parameters =
-Dict{Any, Any}(
-    "source" => Any[
-                    Dict{Any, Any}(
-                        "pwr" => 200e3,
-                        "control_type" => "RL",
-                        "fltr" => "L",
-                        #"L1" => 0.0008,
-                        ),
-                    Dict{Any, Any}(
-                        "pwr" => 200e3,
-                        "fltr" => "LC",
-                        "control_type" => "classic",
-                        "mode" => "Droop",),
-                    ],
-    "grid" => Dict{Any, Any}(
-        "phase" => 3,
-        "ramp_end" => 0.04,)
-)
-
-
-
-function reward_function(env, name = nothing)
-
-    if name == "classic"
-        return 0
-    else
-        p_ref = 1e3
-        q_ref = 0e3
-
-        pq0_ref = [p_ref; q_ref; 0]
-
-        v_a_meas = env.x[findfirst(x -> x == "source1_v_C_cables_a", env.state_ids)]
-        v_b_meas = env.x[findfirst(x -> x == "source1_v_C_cables_b", env.state_ids)]
-        v_c_meas  = env.x[findfirst(x -> x == "source1_v_C_cables_c", env.state_ids)]
-        i_a_meas = env.state[findfirst(x -> x == "source1_i_L1_a", env.state_ids)]
-        i_b_meas = env.state[findfirst(x -> x == "source1_i_L1_b", env.state_ids)]
-        i_c_meas  = env.state[findfirst(x -> x == "source1_i_L1_c", env.state_ids)]
-
-        v_abc_meas = [v_a_meas; v_b_meas; v_c_meas]
-        i_abc_meas = [i_a_meas; i_b_meas; i_c_meas]
-
-        i_abc_ref = InvClarkeTransform(Inv_p_q_v(ClarkeTransform(v_abc_meas), pq0_ref))
-
-        i_abc_ref = clamp.(i_abc_ref/env.nc.parameters["source"][1]["i_limit"], -0.95, 0.95)
-
-        #i_abc_ref= [0; 0; 0]
-
-        #println(i_abc_ref)
-
-        # if any((abs.(env.state)).>=1)
-        #     return -1
-        # end
-
-        r = 1 - 1/3*(sum((abs.(i_abc_ref - i_abc_meas)/2).^0.5))
-
-        r += (0.2 - clamp(maximum(abs.(env.state)), 0.2, 1.0))
-
-        return r * (1 - 0.99)/2
-    end
-
-end
-
-function reference(t)
-    if t < 0.04
-        return [0.0, 0.0, 0.0]
-    end
-
-    θ = 2*pi*50*t
-    θph = [θ; θ - 120π/180; θ + 120π/180]
-    return +10 * cos.(θph) 
-end
-
-featurize_ddpg = function(state, env, name)
-    if name == "ElectricGrid_ddpg_1"
-
-        #state = state[findall(x -> split(x, "_")[2] == "i" , env.agent_dict["ElectricGrid_ddpg_1"]["state_ids"])]
-
[…]
