+++ src/classical_control.jl
@@ -1091,7 +1091,9 @@ function SwingMode(Source::ClassicalControls, num_source, V_filt_cap_new; t_end 
 
     Source.Vd_abc_new[num_source, :, end] = 2*Source.V_ref[num_source, :]/Source.Vdc[num_source]
 
-    PhaseLockedLoop3ph(Source, num_source, V_filt_cap_new)
+    #PhaseLockedLoop3ph(Source, num_source, V_filt_cap_new)
+    Source.fpll, Source.θpll, Source.pll_err_t, Source.pll_err = PhaseLockedLoop3ph(Source.θpll, Source.pll_err, Source.pll_err_t,
+Source.ts, Source.fsys, Source.fpll, num_source, V_filt_cap_new)
 
     Source.f_source[num_source, :, end] = Source.fsys*[1 1 1]
     Source.θ_source[num_source, :, end] = θph
@@ -1137,7 +1139,8 @@ function VoltageControlMode(Source::ClassicalControls, num_source, V_filt_cap_ne
     VoltageController(Source, num_source, θ, ω, V_filt_cap_new, I_filt_poc_new)
     CurrentController(Source, num_source, θ, ω, V_filt_cap_new, I_filt_inv_new)
 
-    PhaseLockedLoop3ph(Source, num_source, V_filt_cap_new)
+    Source.fpll, Source.θpll, Source.pll_err_t, Source.pll_err = PhaseLockedLoop3ph(Source.θpll, Source.pll_err, Source.pll_err_t,
+Source.ts, Source.fsys, Source.fpll, num_source, V_filt_cap_new)
 
     Source.f_source[num_source, :, end] = Source.fsys*[1 1 1]
     Source.θ_source[num_source, :, end] = θph
@@ -1157,14 +1160,15 @@ function DroopControlMode(Source::ClassicalControls, num_source, V_filt_cap_new,
 
     Vrms = Ramp(pu*Source.Vrms[num_source], Source.ts, Source.steps; t_end = t_end)
 
-    DroopControl(Source, num_source, Vrms = Vrms)
+    Source.ω_droop, Source.θ_droop, Source.V_ref = DroopControl(Source.ω_droop, Source.θ_droop, Source.p_q_poc, Source.fsys, Source.D, Source.ts, Vrms, Source.V_ref, num_source)
     θ = Source.θ_droop[num_source, 1]
     ω = Source.ω_droop[num_source, 1, end]
 
     VoltageController(Source, num_source, θ, ω, V_filt_cap_new, I_filt_poc_new)
     CurrentController(Source, num_source, θ, ω, V_filt_cap_new, I_filt_inv_new)
 
-    PhaseLockedLoop3ph(Source, num_source, V_filt_cap_new)
+    Source.fpll, Source.θpll, Source.pll_err_t, Source.pll_err = PhaseLockedLoop3ph(Source.θpll, Source.pll_err, Source.pll_err_t,
+Source.ts, Source.fsys, Source.fpll, num_source, V_filt_cap_new)
 
     Source.f_source[num_source, :, end] = (ω/2π)*[1 1 1]
     Source.θ_source[num_source, :, end] = [θ; θ - 120π/180; θ + 120π/180]
@@ -1184,7 +1188,8 @@ function PQControlMode(Source::ClassicalControls, num_source, pq0, V_filt_cap_ne
         pq0 = pq0.*(Source.S[num_source]/norm(pq0))
     end
 
-    PhaseLockedLoop3ph(Source, num_source, V_filt_cap_new)
[…]
