+++ src/agent_td3.jl
@@ -1,0 +1,202 @@
+using Flux
+
+# TD3 uses two critics 
+struct TD3Critic
+    critic_1::Flux.Chain
+    critic_2::Flux.Chain
+end
+
+# make vertical concatenation of state and action,
+# then feed to critic_1 and critic_2
+Flux.@functor TD3Critic
+(c::TD3Critic)(s, a) = (inp = vcat(s, a); (c.critic_1(inp), c.critic_2(inp)))
+
+mutable struct TD3Policy{
+    BA<:NeuralNetworkApproximator,
+    BC<:NeuralNetworkApproximator,
+    TA<:NeuralNetworkApproximator,
+    TC<:NeuralNetworkApproximator,
+    P,
+    R<:AbstractRNG,
+} <: AbstractPolicy
+
+    behavior_actor::BA
+    behavior_critic::BC
+    target_actor::TA
+    target_critic::TC
+    γ::Float32
+    ρ::Float32
+    batch_size::Int
+    start_steps::Int
+    start_policy::P
+    update_after::Int
+    update_freq::Int
+    policy_freq::Int
+    target_act_limit::Float64
+    target_act_noise::Float64
+    act_limit::Float64
+    act_noise::Float64
+    update_step::Int
+    rng::R
+    replay_counter::Int
+    # for logging
+    actor_loss::Float32
+    critic_loss::Float32
+end
+
+"""
+    TD3Policy(;kwargs...)
+
+# Keyword arguments
+
+- `behavior_actor`,
+- `behavior_critic`,
+- `target_actor`,
+- `target_critic`,
+- `start_policy`,
+- `γ = 0.99f0`,
+- `ρ = 0.995f0`,
+- `batch_size = 32`,
+- `start_steps = 10000`,
+- `update_after = 1000`,
+- `update_freq = 50`,
+- `policy_freq = 2` # frequency in which the actor performs a gradient update_step and critic target is updated
+- `target_act_limit = 1.0`, # noise added to actor target
+- `target_act_noise = 0.1`, # noise added to actor target
+- `act_limit = 1.0`, # noise added when outputing action
+- `act_noise = 0.1`, # noise added when outputing action
+- `update_step = 0`,
+- `rng = Random.default_rng()`,
+"""
+function TD3Policy(;
+    behavior_actor,
+    behavior_critic,
+    target_actor,
+    target_critic,
+    start_policy,
+    γ=0.99f0,
+    ρ=0.995f0,
+    batch_size=64,
+    start_steps=10000,
+    update_after=1000,
+    update_freq=50,
+    policy_freq=2,
+    target_act_limit=1.0,
+    target_act_noise=0.1,
+    act_limit=1.0,
+    act_noise=0.1,
+    update_step=0,
+    rng=Random.default_rng()
+)
+    copyto!(behavior_actor, target_actor)  # force sync
+    copyto!(behavior_critic, target_critic)  # force sync
+    TD3Policy(
+        behavior_actor,
+        behavior_critic,
+        target_actor,
+        target_critic,
+        γ,
+        ρ,
+        batch_size,
+        start_steps,
+        start_policy,
+        update_after,
+        update_freq,
+        policy_freq,
+        target_act_limit,
+        target_act_noise,
+        act_limit,
+        act_noise,
+        update_step,
+        rng,
+        1, # keep track of numbers of replay
+        0.0f0,
+        0.0f0,
+    )
+end
+
+# TODO: handle Training/Testing mode
+function RLBase.plan!(p::TD3Policy, env)
+    p.update_step += 1
+
+    if p.update_step <= p.start_steps
+        p.start_policy(env)
+    else
+        D = device(p.behavior_actor)
+        s = state(env)
[…]
