### 2023-06-22 bda1cd7
#100 add test scripts
