### 2023-06-27 241db01
Add droop example script
Co-authored-by: SeptimusBoshoff <SeptimusBoshoff@users.noreply.github.com>
