+++ examples/scripts/RL_Droop.jl
@@ -1,5 +1,81 @@
 using ElectricGrid
 
+mutable struct PLL
+
+    total_phases::Int64 # always has to be 3 when using pll 3ph
+
+    # initialise matrices
+    θpll::Array{Float64} # two past terms
+    fpll::Array{Float64} # four past terms
+    pll_err_t::Array{Float64}
+    pll_err::Array{Float64}
+
+    Ki::Float64
+    Kp::Float64
+
+    ts::Float64
+    fsys::Float64
+
+    #=     function PLL(total_phases::Int64, θpll::Array{Float64},
+                fpll::Array{Float64}, pll_err_t::Array{Float64}, pll_err::Array{Float64}, Ki::Float64, Kp::Float64, ts::Float64)
+
+        new(total_phases, θpll, fpll, pll_err_t, pll_err, Ki, Kp, ts)
+    end
+    =#
+    function PLL(fsys, ts; ωn = fsys + 20, ξ = 0.35)
+
+        total_phases = 3 # always has to be 3 when using pll 3ph
+
+        # initialise matrices
+        θpll = zeros(total_phases, 2) # two past terms
+        fpll = fsys*ones(total_phases, 4) # four past terms
+        pll_err_t = zeros(total_phases)
+        pll_err = zeros(total_phases, 3)
+
+        # tuning
+        Ki = (ωn)^2 # tuning -> to struct
+        Kp = (ξ)*2*sqrt(Ki) # tuning
+
+        #PLL(total_phases, θpll, fpll, pll_err_t, pll_err, Ki, Kp, ts)
+        new(total_phases, θpll, fpll, pll_err_t, pll_err, Ki, Kp, ts, fsys)
+    end
+
+    function (pll::PLL)(V_filt_cap_new)
+
+        θ = pll.θpll[1, end]
+
+        v_αβγ = ClarkeTransform(V_filt_cap_new)
+
+        if norm(v_αβγ) != 0
+            v_αβγ = sqrt(3)*v_αβγ./norm(v_αβγ)
+        end
+
+        err_new = v_αβγ[2]*cos(θ) - v_αβγ[1]*sin(θ)
+
+        f_new, err_t_new, err_int =
+        PIController(err_new, pll.pll_err[1, :], pll.pll_err_t[1], pll.Kp, pll.Ki, pll.ts, bias = pll.fsys, max_t_err = 0.00015)
+
+        θ = ThirdOrderIntegrator(θ, pll.ts, 2π*[pll.fpll[1, 2:end]; f_new])
+
+        roll(pll.fpll)
+        pll.fpll[1, end] = f_new[1]
+        pll.fpll[2, end] = f_new[1]
+        pll.fpll[3, end] = f_new[1]
+
+        roll(pll.θpll)
+        pll.θpll[1, end] = (θ.%(2π))
+        pll.θpll[2, end] = ((θ - 120π/180).%(2π))
+        pll.θpll[3, end] = ((θ + 120π/180).%(2π))
+
+        pll.pll_err_t[1] = err_t_new[1]
+        pll.pll_err[1, :] = err_int
+
+        return fpll, θpll
+    end
+end
+
+println("...........o0o----ooo0§0ooo~~~  START  ~~~ooo0§0ooo----o0o...........\n\n")
+
 CM = [ 0. 0. 1.
         0. 0. 2.
         -1. -2. 0.]
@@ -26,17 +102,59 @@ Dict{Any, Any}(
         "ramp_end" => 0.04,)
 )
 
+#-------------------------------------------------------------------------------
+total_num_RL = 1
+total_phases = 3 # always has to be 3 when using pll 3ph
+
+# initialise matrices
[…]
