+++ examples/scripts/test_nodec_non_linear.jl
@@ -1,35 +1,21 @@
 using ElectricGrid
 using LinearAlgebra
 using ControlSystemsBase
-
-function checkifnonlinear(list::Matrix)
-    return checkifnonlinear([list])
-end
-
-function checkifnonlinear(list::Array)
-    for element in list
-        for idx in eachindex(element)
-            if typeof(element[idx]) <: Function
-                return true
-            end
-        end
-    end
-    return false
-end
-
-
+using BenchmarkTools
 
 CM = [0 1
     -1 0]
-    
+
+# start: Setting up parameters
 # Source
 R = 1.1e-3
 L = 70e-6
 C = 250e-6
 
+LL(i) = L_low_1 + (L_high_1-L_low_1)/2 * (1-2/π*atan(i)) + L_low_2 + (L_high_2-L_low_2)/2*(1-2/π*atan(i))
 # Cable
-C_b = 1e-4/2
-L_b = x->1e-4
+C_b = 1e-4 / 2
+L_b = x -> 1e-4
 R_b = 1e-3
 
 # Load
@@ -40,7 +26,7 @@ L_l = 1e-2;
 parameters = Dict()
 
 grid_properties = Dict()
-grid_properties["fs"] =  10e3
+grid_properties["fs"] = 10e3
 grid_properties["v_rms"] = 230
 grid_properties["phase"] = 1;
 parameters["grid"] = grid_properties
@@ -48,11 +34,9 @@ parameters["grid"] = grid_properties
 source1 = Dict()
 source_list = []
 
-# source1["fltr"] = "LC"
 source1["fltr"] = "L"
 source1["R1"] = R
 source1["L1"] = L
-# source1["C"] = C
 
 push!(source_list, source1)
 
@@ -78,24 +62,81 @@ load1["C"] = C_l;
 push!(load_list, load1);
 parameters["load"] = load_list;
 
+# end: Setting up parameters
+
 nc = NodeConstructor(num_sources=1, num_loads=1, parameters=parameters, CM=CM);
 
-A, B, F, D = GetSystem(nc);
+A, B, C, D = GetSystem(nc);
 
-@show checkifnonlinear(A)
+# (rows, columns) = size(A)
+# AA = Matrix{Any}(undef, (rows, columns))
+# for row in 1:rows
+#     for column in 1:columns
+#         h = A[row, column]
+#         if isa(h, Number)
+#             AA[row, column] = x -> h
+#         else
+#             AA[row, column] = h
+#         end
+#     end
+# end
 
-(rows, columns) = size(A)
-AA = Matrix{Any}(undef, (rows, columns))
-for row in 1:rows
-    for column in 1:columns
-        h = A[row, column]
-        if isa(h, Number)
-            AA[row, column] = x -> h
-        else
-            AA[row, column] = h
+# # Here we could try to make it much faster for example
+# # when we remember in which line functions are and in which line not. (Pipe operator)
+# AAA(x) = (|>).(x, AA)
+
+# x_state_space = [0 0 0 0 0]
+# AAAA = AAA(x_state_space)
+
+
+function get_b!(A)
+    b = []
+    (col,row) = size(A)
+    for i = 1:col
+        a = A[i,:]
+        if any(x->isa(x,Function)==true,a)
+            append!(b,true)
+            for j = 1:row
+                a_j = A[i,j]
+                if isa(a_j,Number)
+                    A[i,j] = x->a_j
+                end
+            end
+        else append!(b,false)
         end
     end
+    return b
 end
 
-AAA(x) = (|>).(x, AA)
-Z = AAA([0, 0, 0, 0, 0])
+b = get_b!(A)
[…]
