+++ src/electric_grid_env.jl
@@ -1,7 +1,8 @@
+using DifferentialEquations
 mutable struct ElectricGridEnv <: AbstractEnv
     verbosity
     nc
-    sys_d
+    sys
     action_space
     state_space
     done
@@ -33,9 +34,23 @@ mutable struct ElectricGridEnv <: AbstractEnv
     state_parameters
     y #holds all of the inductor voltages and capacitor currents
     agent_dict
+    nonlinear
 end
 
+function checkifnonlinear(list::Matrix)
+    return checkifnonlinear([list])
+end
 
+function checkifnonlinear(list::Array)
+    for element in list
+        for idx in eachindex(element)
+            if typeof(element[idx]) <: Function
+                return true
+            end
+        end
+    end
+    return false
+end
 """
     ElectricGridEnv(...)
 
@@ -91,21 +106,6 @@ https://juliareinforcementlearning.org/
 
 
 """
-function checkifnonlinear(list::Matrix)
-    return checkifnonlinear([list])
-end
-
-function checkifnonlinear(list::Array)
-    for element in list
-        for idx in eachindex(element)
-            if typeof(element[idx]) <: Function
-                return true
-            end
-        end
-    end
-    return false
-end
-
 function ElectricGridEnv(;
     maxsteps=500,
     ts=1 / 10_000,
@@ -129,7 +129,8 @@ function ElectricGridEnv(;
     action_delay=1,
     t_end=nothing,
     verbosity=0,
-    agent_dict=nothing
+    agent_dict=nothing,
+    nonlinear=false
 )
 
     if !(isnothing(t_end))
@@ -172,31 +173,44 @@ function ElectricGridEnv(;
     A, B, C, D = GetSystem(nc)
     #XXX Here is where the Linear or nonlinear solver should be
 
-    checkifnonlinear([A,B,C,D])
-    if checkifnonlinear([A,B,C,D])
-        
-    else
-
+    if checkifnonlinear([A, B, C, D])
+        nonlinear = true
     end
 
-    Ad = exp(A * ts) #fastExpm(A*ts) might be a better option
-    #Bd = A \ (Ad - I) * B #This may be bad for large sizes, maybe QR factorise, then use ldiv!
-    Bd = (Ad - I) * B #
-    ldiv!(factorize(A), Bd)
-    sys_d = HeteroStateSpace(Ad, Bd, C, D, Float64(ts))
-    state_parameters = GetStateParameters(nc)
+    if nonlinear
+        sys = [A, B, C, D]
+        if use_gpu
+            A = CuArray(A)
+            B = CuArray(B)
+            C = CuArray(C)
+            if isa(D, Array)
+                D = CuArray(D)
+            end
+        end
+    else
+        AA = Array{Float64}(A)
+        BB = Array{Float64}(B)
+        CC = Array{Float64}(C)
+        Ad = exp(AA * ts) #fastExpm(A*ts) might be a better option
+        Bd = AA \ (Ad - I) * BB #This may be bad for large sizes, maybe QR factorise, then use ldiv!
+        # Bd = (Ad - I) * B #
+        # ldiv!(factorize(Ad), Bd)
+        sys = HeteroStateSpace(Ad, Bd, CC, D, Float64(ts))
 
-    if use_gpu
-        Ad = CuArray(A)
-        Bd = CuArray(B)
-        C = CuArray(C)
-        if isa(D, Array)
-            D = CuArray(D)
+        if use_gpu
+            Ad = CuArray(A)
+            Bd = CuArray(B)
+            C = CuArray(C)
+            if isa(D, Array)
+                D = CuArray(D)
+            end
         end
     end
 
[…]
