+++ Project.toml
@@ -3,6 +3,7 @@ uuid = "d6749236-802c-4af3-9b09-2fcebf7269d7"
 version = "1.0.0"
 
 [deps]
+BenchmarkTools = "6e4b80f9-dd63-53aa-95a3-0cdb28fa8baf"
 CDDLib = "3391f64e-dcde-5f30-b752-e11513730f60"
 CSV = "336ed68f-0bac-5ca0-87d4-7b16caf5d00b"
 CUDA = "052768ef-5323-5732-b1bb-66c8b64840ba"

[…]
