### 2023-06-30 #5424 @osorensen
:wave: @kiranshila, @degleris1, @matt-graham would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
