### 2023-07-01 upb-lea/ElectricGrid.jl#103 @osorensen
Pre-review editorial comments on JOSS submission
This is related to the following JOSS submission, https://github.com/openjournals/joss-reviews/issues/5424.

The figure captions don't appear as they should, see example below:

Before we start the review, please modify the paper so the figure captions follow our style guide. That is, `![Caption for example figure.\label{fig:example}](figure.png)`. See details [here](https://joss.readthedocs.io/en/latest/submitting.html).

For this particular example, the change should be from

```
![\label{fig:jeg}](Overview_EG.png)
_Fig. 2:  Overview of the functionality and interconnections of the ElectricGrid.jl framework._
```

to

```
![Overview of the functionality and interconnections of the ElectricGrid.jl framework.\label{fig:label}](Overview_EG.png)
```
