### 2023-07-01 upb-lea/ElectricGrid.jl#105 @osorensen
Pre-review editorial comment on JOSS submission
I believe some of these links point to the wrong URL:

In particular, the last one named "Training an RL agent on the electrical grid" points to [here](https://github.com/upb-lea/ElectricGrid.jl/blob/main/examples/notebooks/NodeConstructor_Theory_DEMO.ipynb). Is that correct?

Please double check all of them by clicking the links in the generated pdf.

https://github.com/openjournals/joss-reviews/issues/5424#issuecomment-1615004811
