### 2023-07-01 upb-lea/ElectricGrid.jl#106 @osorensen
Pre-review editorial comment on JOSS submission
Please spell Julia with uppercase first consistently. Here is an example:

https://github.com/openjournals/joss-reviews/issues/5424#issuecomment-1615004811
