### 2023-07-02 b8b7e89
Pre-review editorial comments on JOSS submission #103
changed figure captions according to style guide
