### 2023-07-02 ae0238c
Pre-review editorial comment on JOSS submission #105
corrected RL example link
changed classic controller link from auxiliaries to droop
