### 2023-07-02 488e157
Pre-review editorial comment on JOSS submission #106
corrected spelling julia - Julia
