### 2023-07-02 cf9c71b
Suggested change to JOSS submission (#104)
* Citation should be in-text, not parenthesized.

* Capitalize Jupyter Notebook
