+++ examples/scripts/env_example_nonlinear.jl
@@ -1,62 +1,125 @@
 using ElectricGrid
 using DifferentialEquations
 using PlotlyJS
-#using ReinforcementLearning
-CM = [0. 1.
-     -1. 0.]
 
-S_source = 1e4
+CM = [0 1
+    -1 0]
 
-S_load = 1e2
-pf_load = 1
-v_rms = 230
-R_load, L_load, X, Z = ParallelLoadImpedance(S_load, pf_load, v_rms)
+# Source
+R = 1.1e-3
+L = 70e-6
+C = 250e-6
 
-parameters = Dict{Any, Any}(
-    "source" => Any[
-                    Dict{Any, Any}("pwr" => S_source, "control_type" => "classic", "mode" => "Swing", "fltr" => "LC", "i_limit" => 1e4, "v_limit" => 1e4,),
-                    ],
-    "load"   => Any[
-                    Dict{Any, Any}("impedance" => "RLC", "R" => R_load, "v_limit" => 1e4, "i_limit" => 1e4)
-                    ],
-    "cable"   => Any[
-                    Dict{Any, Any}("R" => 1e-3, "L" => x->1e-4, "C" => 1e-4),
-                    ],
-    "grid" => Dict{Any, Any}("fs"=>1e4, "phase"=>3, "v_rms"=>230, "f_grid" => 50, "ramp_end"=>0.00)
-)
+# Cable
+C_b = 1e-4/2
+L_b = 1e-4
+R_b = 1e-3
 
-env1 = ElectricGridEnv(CM = CM, parameters = parameters, verbosity = 2)
+# Load
+R_l = 100
+C_l = 1e-2
+L_l = 1e-2;
+parameters = Dict()
 
-parameters = Dict{Any, Any}(
-    "source" => Any[
-                    Dict{Any, Any}("pwr" => S_source, "control_type" => "classic", "mode" => "Swing", "fltr" => "LC", "i_limit" => 1e4, "v_limit" => 1e4,),
-                    ],
-    "load"   => Any[
-                    Dict{Any, Any}("impedance" => "RLC", "R" => R_load, "v_limit" => 1e4, "i_limit" => 1e4)
-                    ],
-    "cable"   => Any[
-                    Dict{Any, Any}("R" => 1e-3, "L" => 1e-4, "C" => 1e-4),
-                    ],
-    "grid" => Dict{Any, Any}("fs"=>1e4, "phase"=>3, "v_rms"=>230, "f_grid" => 50, "ramp_end"=>0.00)
-)
+grid_properties = Dict()
+grid_properties["fs"] =  10e3
+grid_properties["v_rms"] = 230
+grid_properties["phase"] = 1;
+parameters["grid"] = grid_properties
 
+source1 = Dict()
+source_list = []
 
-env2 = ElectricGridEnv(CM = CM, parameters = parameters, verbosity = 2)
-value = 6
+source1["fltr"] = "L"
+source1["R1"] = R
+source1["L1"] = L
+
+push!(source_list, source1)
+
+parameters["source"] = source_list
+
+cable = Dict()
+cable["R"] = R_b
+cable["L"] = L_b
+cable["C"] = C_b
+cable_list = []
+
+push!(cable_list, cable);
+parameters["cable"] = cable_list
+
+load1 = Dict()
+load_list = []
+
+load1["impedance"] = "RLC"
+load1["R"] = R_l;
+load1["L"] = L_l;
+load1["C"] = C_l;
+
+push!(load_list, load1);
+parameters["load"] = load_list;
+
+# @show parameters
[…]
