### 2023-07-03 290288e
Joss (#109)
* Pre-review editorial comments on JOSS submission #103

changed figure captions according to style guide

* Pre-review editorial comment on JOSS submission #105

corrected RL example link
changed classic controller link from auxiliaries to droop

* Pre-review editorial comment on JOSS submission #106

corrected spelling julia - Julia
