+++ src/ElectricGrid.jl
@@ -33,6 +33,7 @@ include("./custom_control.jl")
 include("./solar_module.jl")
 include("./electric_grid_env.jl")
 include("./agent_ddpg.jl")
+include("./agent_td3.jl")
 include("./classical_control.jl")
 include("./multi_controller.jl")
 include("./render.jl")
