+++ src/agent_ddpg.jl
@@ -13,7 +13,7 @@ using MacroTools: @forward
 (policy::AbstractPolicy)(stage::AbstractStage, env::AbstractEnv, action, training::Bool) = policy(stage, env, action)
 
 
-#re-definition needed to use ElectricGrid-internal action-space functions
+# re-definition needed to use ElectricGrid-internal action-space functions
 function RLBase.update!(
         trajectory::AbstractTrajectory,
         policy::AbstractPolicy,
@@ -233,7 +233,6 @@ end
 
 
 # also in a sep src
-
 global rngg = StableRNG(123)
 global initt = Flux.glorot_uniform(rngg)
 
