### 2023-07-06 d029814
#100 log first results with TD3
