### 2023-07-07 b7d6dfa
Changed the accuracy of the ODE solver to speed up
