+++ examples/scripts/env_example_nonlinear.jl
@@ -12,7 +12,14 @@ v_rms = 230
 R_load, L_load, X, Z = ParallelLoadImpedance(S_load, pf_load, v_rms)
 
 # Which value is plottet
-value = 2
+value = 4
+function NonlinearInductance(drop, length)
+    # drop, is to how much it decreses with i to infty
+    # length, length is the point, where its half decresed
+    return x->(1-(drop/100)*1/(1+exp(-(abs(x)-length)/2)))
+end
+
+l = NonlinearInductance(90,1)
 
 parameters = Dict{Any, Any}(
     "source" => Any[
@@ -22,20 +29,20 @@ parameters = Dict{Any, Any}(
                     Dict{Any, Any}("impedance" => "R", "R" => R_load, "v_limit" => 1e4, "i_limit" => 1e4)
                     ],
     "cable"   => Any[
-                    Dict{Any, Any}("R" => 1e-3, "L" => x->1e-4, "C" => 1e-4),
+                    Dict{Any, Any}("R" => 1e-3, "L" => x->1e-4*l(x), "C" => 1e-4),
                     ],
     "grid" => Dict{Any, Any}("fs"=>1e4, "phase"=>3, "v_rms"=>230, "f_grid" => 50, "ramp_end"=>0.00)
 )
 
 env1 = ElectricGridEnv(CM = CM, parameters = parameters, verbosity = 2)
 
-sol = []
+sol_1 = []
 for i = 1:1000
     env1([0.5, 0.5, 0.5])
-    append!(sol,env1.x[value])
+    append!(sol_1,env1.x[value])
 end
 t_t = collect(env1.t0:env1.ts:env1.t)
-p1 = scatter(x=t_t,y=sol,mode="lines",name="nonlinear")
+p1 = scatter(x=t_t,y=sol_1,mode="lines",name="nonlinear")
 
 parameters2 = Dict{Any, Any}(
     "source" => Any[
@@ -52,12 +59,12 @@ parameters2 = Dict{Any, Any}(
 
 env2 = ElectricGridEnv(CM = CM, parameters = parameters2, verbosity = 2)
 
-sol = []
+sol_2 = []
 for i = 1:1000
     env2([0.5, 0.5, 0.5])
-    append!(sol,env2.x[value])
+    append!(sol_2,env2.x[value])
 end
 t_t = collect(env2.t0:env2.ts:env2.t)
-p2 = scatter(x=t_t,y=sol,mode="lines",name="linear")
+p2 = scatter(x=t_t,y=sol_2,mode="lines",name="linear")
 
 plot([p1,p2])
