+++ src/custom_control.jl
@@ -61,7 +61,7 @@ function CustomNonlinearsim(AAA,B,u,tspan,x0)
 
     prob = ODEProblem(f, x0, tspan, u)
     alg = Tsit5()
-    sol = solve(prob, alg, reltol=1e-10, abstol=1e-10)
+    sol = solve(prob, alg, reltol=1e-5, abstol=1e-5)
     xout = sol.u[end]
 
     xout_float = Array{Float64}(undef,0)
