### 2023-07-13 557c32c
Safety in RL #101 started with condensed matrices
