### 2023-07-14 d33c47a
amplitude bugfix, osc3 mode, sanity check
