### 2023-07-14 3d430a3
Adjusting the notebook and src to the current version.
