### 2023-07-14 dc1e46e
Safety in RL #101 added condensed matrices
