### 2023-07-18 c58f164
Changed function within PV arrays. Added battery model. Changed DEMO PV to DC link models.
