+++ examples/scripts/Safeguard_test.jl
@@ -1,31 +1,29 @@
 using ElectricGrid
 using LinearAlgebra
-
-CM = [ 0. 0. 1.
-        0. 0. 2.
-        -1. -2. 0.]
-
-#R_load, L_load, _, _ = ParallelLoadImpedance(50e3, 0.95, 230)
+using Polyhedra
 
 parameters =
 Dict{Any, Any}(
     "source" => Any[
                     Dict{Any, Any}(
-                        "pwr" => 200e3,
+                        "pwr" => 10e3,
                         "fltr" => "LCL",
                         "L1" => 70e-6,
                         "R1" => 1.1e-3,
                         "C" => 300e-6,
                         "R_C" => 7e-3,
                         "control_type" => "classic",
-                        "mode" => "Droop",),
+                        "mode" => "Droop",
+                        "v_limit" => 60,
+                        "i_limit" => 40,
+                        "vdc" => 50),
                     ],
     "load"   => Any[
         Dict{Any, Any}(
             "impedance" => "R",
             "R" => 100,
-            "v_limit" => 1e4,
-            "i_limit" => 1e4)
+            "v_limit" => 1e2,
+            "i_limit" => 1e2)
         ],
     "grid" => Dict{Any, Any}(
         "phase" => 1,
@@ -35,7 +33,6 @@ Dict{Any, Any}(
 
 
 env = ElectricGridEnv(
-    #CM =  CM,
     parameters = parameters,
     t_end = 1,
     action_delay = 0,
@@ -55,15 +52,33 @@ env.A
     - Same?
  - Use it to ensure safety
 =#
+#=
+A = env.A[1:2,1:2]
+B = env.B[1:2]
+=#
 
-A_P10 = env.A[1:2,1:2]
-B_P10 = env.B[1:2]
+# Use matlab vals to compare
+A = [-115.705536  -14284.4644;
+     3333.04169 -41.6630212]
+B = [14285.7143; 0 ]
+
+A_P10 = exp(A * env.ts) #fastExpm(A*ts) might be a better option
+    #Bd = A \ (Ad - I) * B #This may be bad for large sizes, maybe QR factorise, then use ldiv!
+B_P10 = A \ (A_P10 - I) * B #
+
+
+i_dead = env.nc.parameters["source"][1]["i_limit"]
+v_dead = env.nc.parameters["source"][1]["v_limit"]
+v_dc = env.nc.parameters["source"][1]["vdc"]
+A_P10[1,2] = A_P10[1,2]*v_dead/i_dead;
+A_P10[2,1] = A_P10[2,1]*i_dead/v_dead;
+B_P10[1,1] = B_P10[1,1]/i_dead*v_dc;
+B_P10[2,1] = B_P10[2,1]*v_dc/v_dead;
+
 
 W_u = [-1; 1]
-omega_u = [
-    env.nc.parameters["source"][1]["vdc"];
-    env.nc.parameters["source"][1]["vdc"]
-         ]
+
+omega_u = [1; 1]
 
 
 W_x = [
@@ -73,12 +88,66 @@ W_x = [
     0 1
     ]
 
+v_ref = 40
+v_lim_poly = 1.2 * v_ref
+i_lim_poly = 30
+
+#=
 omega_x = [
     env.nc.parameters["source"][1]["i_limit"];
     env.nc.parameters["source"][1]["i_limit"];
     env.nc.parameters["source"][1]["v_limit"];
[…]
