+++ Project.toml
@@ -29,6 +29,8 @@ Logging = "56ddb016-857b-54e1-b83d-db4d58db5568"
 LoggingExtras = "e6f89c97-d47a-5376-807f-9c37f3926c36"
 MAT = "23992714-dd62-5051-b70f-ba57cb901cac"
 MacroTools = "1914dd2f-81c6-5fcd-8719-6d5c9610ff09"
+Makie = "ee78f7c6-11fb-53f2-987a-cfe4a2b5a57a"
+MeshCat = "283c5d60-a78f-5afe-a0af-af636b173e11"
 NLSolvers = "337daf1e-9722-11e9-073e-8b9effe078ba"
 NonlinearSolve = "8913a72c-1f9b-4ce2-8d82-65094dcecaec"
 PlotlyBase = "a03496cd-edff-5a9b-9e67-9cda94a718b5"
