### 2023-07-21 1d5459c
Safety in RL #101 - shifted polytope calculation functions
