+++ examples/scripts/Safeguard.jl
@@ -1,0 +1,257 @@
+using ElectricGrid
+using LinearAlgebra
+using Polyhedra
+
+############################################################################################
+# New functions
+
+"""
+    my_blkdiag_from_vec(v, N)
+
+# Description
+Creates a blockdiagonal matrix with N-times vector v on the diagonal
+
+# Arguments
+
+# Keyword Arguments
+- `v::Vector`: Vector that appears on diagonal.
+- `N::Int`: number of diagonal elements (v).
+
+# Return Values
+- `M_blk::Matrix`: Blockdiagonal Matrix with vector v on diagonal
+
+"""
+function my_blkdiag_from_vec(v, N)
+
+    count = 1
+    v_lec = length(v)
+    M_blk = zeros(v_lec*N, N)
+
+    for n in 1:N+1
+        for i in 1:size(v)[1]+1
+            if n == i
+                M_blk[count, i] = v[1]
+                M_blk[count+v_lec-1, i] = v[v_lec]
+            end
+        end
+        count+=v_lec
+    end
+
+    return M_blk
+end
+
+"""
+    my_blkdiag_matrix(m, N)
+
+# Description
+Creates a blockdiagonal matrix with N-times matrix m on the diagonal
+
+# Arguments
+
+# Keyword Arguments
+- `m::Matrix`: Vector that appears on diagonal.
+- `N::Int`: number of diagonal elements (v).
+
+# Return Values
+- `M_blk::Matrix`: Blockdiagonal Matrix with vector v on diagonal
+"""
+function my_blkdiag_matrix(m, N)
+
+    m_lines, m_rows = size(m)
+    M_blk = zeros(m_lines*(N+1), m_rows*(N+1))
+    for n in 1:N+1
+        for line in 1:(m_lines)
+            for row in 1:(m_rows)
+                M_blk[line+m_lines*(n-1), row+m_rows*(n-1)] = m[line, row]
+            end
+        end
+    end
+    return M_blk
+end
+
+"""
+    get_condensed_matrices(A, B, omega_x, omega_u, W_x, W_u, N)
+
+# Description
+Calulates the condesed state space and contraints matrices
+
+# Arguments
+
+# Keyword Arguments
+- `A::Matrix`: .
+- `B::Matrix`: .
+
+# Return Values
+-
+"""
+function get_condensed_matrices(A, B, omega_x, omega_u, W_x, W_u, N)
+
+    m = 1
+    n = size(B)[1]
+
+    A_N = 1* Matrix(I, size(A)[1], size(A)[1])
+    B_N = zeros(n, m*N)
+
+    Omega_x = omega_x
+    Omega_u = []
+
+
+    W_x_cal = my_blkdiag_matrix(W_x, N)
+    W_u_cal = my_blkdiag_from_vec(W_u, N)
+
+    for ii  = 1:N
+
+        A_N = [A_N; A_P10^ii]
+        Omega_u = [Omega_u; omega_u]
+        Omega_x = [Omega_x; omega_x]
+
+        B_N_newline = A^(ii-1)*B
+
+
+        for jj in 2:ii
+
+            B_N_newline = hcat(B_N_newline, A^(ii-jj)*B)
+
+
+        end
+        B_N_newline = hcat(B_N_newline, zeros(n, m*(N-ii)))
+        B_N = [B_N; B_N_newline]
+
+    end
+
+    return A_N, B_N, Omega_x, Omega_u, W_x_cal, W_u_cal
+end
+
[…]
