### 2023-07-23 #5616 @degleris1
@osorensen @Webbah I have completed my review. I think some minor things should be addressed, but nothing major.

General Comments
- The statement of need is missing from both the paper and the documentation.
- How does this software package compare to [PowerSimulationDynamics.jl](https://github.com/NREL-Sienna/PowerSimulationsDynamics.jl)? The "state of field" section in the paper seems to miss this package, which would seem to have similar functionality (dynamic, not quasi-steady state, power system simulation). This seems like something worth explaining in the paper.

Small Things
- Paper Line 48 - "quasi-stationar" --> "quasi-stationary"
- Documentation - Update welcome page with correct installation instructions
- (This is just a question/suggestion, not a review item.) Why are parameters untyped nested dictionaries? Seems like it would make much more sense to take advantage of Julia's type system and do `source => [Source(pwr=200e3, mode="VSG"), ...]`.
