### 2023-07-24 #5616 @kiranshila
@osorensen and @Webbah, I've completed my checklist and review, details below.

The software seems to work as advertised and fills a need specifically in the Julia ecosystem. My remaining questions involve how the software compares with state-of-the-art. From a software perspective, my notes are mainly for future improvement as the functionality is complete. As a library consumer, however, I would probably find difficulty in using this package due to the type issues described below. There may be performance consequences, but with no point of reference I'm not sure how to compare and the authors don't make any concrete claims as such.

## Review

### Checklist

- While not explicitly titled, `Statement of need`, the section on `Background on electric grids and their control` describes the problems the software is designed to solve ("demand for comprehensive testing of new control concepts during their development phase") but does not include who is the target audience and what is it's relation to other work. A bit more elaboration on who would use such simulation software (power systems researchers, working power systems engineers, etc.) and why the existing tools don't solve this problem is needed.
- I'm suspicious of the single point of reference in the `State of the field` section, specifically as it references another Julia package. As Julia is still quite a new language, I would assume that the majority of power system engineers use the likes of MATLAB or Python. For example, a quick search shows the likes of [ANDES](https://docs.andes.app/en/latest/), [PyPSA](https://pypsa.readthedocs.io/en/latest/), and [OpenIPSL](https://github.com/OpenIPSL/OpenIPSL). I would like to know what the value-add of Julia is over these options and how your package compares.

### Software

1. I'm not sure if Markdown is the most appropriate way to serialize the `ElectricGridEnv`, at least in a terminal context. You can customize this using `::MIME"text/plain"` in the `show` method. I'd expect many people to interact with this software in the REPL, which the markdown format may be unnecessary.

2. More custom types may benefit from a `show` implementation (like `NodeConstructor`).

3. In several places (dict keys, modes, parameters, etc) strings are used instead of a more specific data structure. For a Julia program, this is a big code smell problem. This will hurt performance as the lack of concrete types will cause the JIT to fall back to RTTI. Additionally, this implies that an invalid state could be representable, which would require robust runtime error checking, which we want to avoid if we can.

For the case of fixed sets of parameters, structs or enums would be more appropriate. For example in the documentation section on [Setting parameters](https://upb-lea.github.io/ElectricGrid.jl/dev/Env_Create/#Setting-Parameters), `As shown above, there are 4 keys: "grid", "source", "cable" and "load".`, yet the types don't prevent me from doing something like

```julia
push!(env.nc.parameters, "Foo"=>"bar")
```

We would always want to prefer explicit over implicit, so something like

```julia
# Concrete field from a struct, named plural as it is a vector
# env.nc.parameters.sources::Vector{Source}

@enum SourceMode swing pq droop synchronverter

struct Source
    L1::Float64
    C::Float64
    mode::SourceMode
    ...
end
```

would be significantly better and probably more performant. This also has a documentation benefit as I'm not sure if every parameter is documented. If I see

```julia
julia> env.nc.parameters["source"][1]
Dict{Any, Any} with 30 entries:
  "L1"           => 0.00170766
  "C"            => 3.08246e-5
  "mode"         => "Synchronverter"
  "fltr"         => "LCL"
    ...
  "X₀"           => 0
```

Some of these parameters I can figure out (or I have found in the documentation), but others like `X₀` I'm not sure how to find. This could be field-specific knowledge that I don't know, however. But, if the set of source parameters were a struct, you can include per-field documentation so there's no ambiguity.

```julia
"""Source Parameters for ..."""
struct SourceParams
    """Inductance (nH) of ..."""
    L1
    ...
end
```

which could then be queried directly in the help REPL (and auto-documented) by

```julia
help?> SourceParams.L1
Inductance (nH) of ...
```

Lastly, for cases where there is a fixed set of string-like things we need to compare, symbols would be appropriate (and canonical in Julia). Strings are always heap-allocated, which is a perf hit.

### Paper

I think the paper may be a bit on the short side. A little more background on why reinforcement learning is appropriate would be welcome (coming from the perspective of a EE with not much knowledge of RL).

Minor notes for readability, feel free to ignore:
- Line 9, and a few others, perhaps `electric grid model` should be used instead of `electric grid` as a naive reader may assume that the grid control itself can be initialized in Julia.
- Lines 24-26, the subject `Power electronics` is used a little strangely here, it seems the electronics themselves describe the application, which seems incorrect. Perhaps `The field of power electronics covers the application...`
