### 2023-07-26 3ab5acf
Safety in RL #101 similar problem with projection
