### 2023-07-28 8ab3d99
JOSS paper #44
Included reviewer comments into the paper

Co-authored-by: MarvinMeyer <MarvinMeyer@users.noreply.github.com>
