+++ docs/JOSS/paper.bib
@@ -55,3 +55,74 @@
   pages={1437--1480},
   year={2015}
 }
+
+@article{ZhangRL2018,
+  author={Zhang, Dongxia and Han, Xiaoqing and Deng, Chunyu},
+  journal={CSEE Journal of Power and Energy Systems}, 
+  title={Review on the research and practice of deep learning and reinforcement learning in smart grids}, 
+  year={2018},
+  volume={4},
+  number={3},
+  pages={362-370},
+  doi={10.17775/CSEEJPES.2018.00520}
+}
+
+@article{GlavicRL2018,
+  title = {Reinforcement Learning for Electric Power System Decision and Control: Past Considerations and Perspectives},
+  journal = {IFAC-PapersOnLine},
+  volume = {50},
+  year = {2017},
+  author = {Mevludin Glavic and Raphaël Fonteneau and Damien Ernst},
+  keywords = {Electric power system, reinforcement learning, control, decision}
+  }
+
+@Article{Brown2017,
+  author  = {Brown, Tom and H{\"o}rsch, Jonas and Schlachtberger, David},
+  journal = {arXiv preprint arXiv:1707.09913},
+  title   = {PyPSA: Python for power system analysis},
+  year    = {2017},
+}
+
+@Article{Thurner2018,
+  author    = {Thurner, Leon and Scheidler, Alexander and Sch{\"a}fer, Florian and Menke, Jan-Hendrik and Dollichon, Julian and Meier, Friederike and Meinecke, Steffen and Braun, Martin},
+  journal   = {IEEE Transactions on Power Systems},
+  title     = {pandapower—an open-source python tool for convenient modeling, analysis, and optimization of electric power systems},
+  year      = {2018},
+  number    = {6},
+  pages     = {6510--6521},
+  volume    = {33},
+  publisher = {IEEE},
+}
+
+   @article{lara2023revisiting,
+    title={Revisiting Power Systems Time-domain Simulation Methods and Models},
+    author={Lara, Jose Daniel and Henriquez-Auba, Rodrigo and Ramasubramanian, Deepak and Dhople, Sairaj and Callaway, Duncan S and Sanders, Seth},
+    journal={arXiv preprint arXiv:2301.10043},
+    year={2023}
+}
+
+@Misc{FaramaFoundation2023,
+  author  = {Farama-Foundation},
+  title   = {Farama-Foundation/Gymnasium: A standard API for single-agent reinforcement learning environments, with popular reference environments and related utilities (formerly Gym)},
+  year    = {2023},
+  journal = {GitHub},
+  url     = {https://github.com/Farama-Foundation/Gymnasium},
+}
+
+@Misc{Donnot2020,
+  author       = {B. Donnot},
+  howpublished = {\url{https://GitHub.com/rte-france/grid2op}},
+  title        = {{Grid2op- A testbed platform to model sequential decision making in power systems. }},
+  year         = {2020},
+  journal      = {GitHub repository},
+  publisher    = {GitHub},
+}
+
+@Misc{Marot2020,
+  author       = {A. Marot and N. Megel and V. Renault and M. Jothy},
+  howpublished = {\url{https://github.com/BDonnot/ChroniX2Grid}},
+  title        = {{ChroniX2Grid - The Extensive PowerGrid Time-serie Generator}},
+  year         = {2020},
+  journal      = {GitHub repository},
+  publisher    = {GitHub},
+}
