### 2023-07-31 7e588ef
Commiting current state of the notebook. Implementation to Env is still missing.
