+++ docs/JOSS/paper.md
@@ -95,9 +95,10 @@ in a (quasi)-stationary state. Also these frameworks tend to focus on large-scal
 which does not allow an evaluation of, e.g., control on component level in case of load fluctuations in the grid.
 In addition, few tools like [@lara2023revisiting] already exist in Julia, which offer dynamic simulations. 
 However, the latter is based on different simplifications (e.g. assumption of a symmetric grid, fixed frequency, ...) and 
-also do not offer an interface to RL toolboxes. In contrast to this, to ensure seamless integration of the control algorithms, 
-the Gymnasium-based API [@FaramaFoundation2023] is implemented in ``ElectricGrid.jl``. For other projects implementing the Gymnasium-based API, such as GridAlive (based on Grid2Op [@Donnot2020]) and ChroniX2Grid [@Marot2020], the focus remains on top-down control and steady-state models.
-Therefore, this contribution provides a tool to close this gap. 
+also do not offer an interface to RL toolboxes.
+To ensure a seamless integration of the control algorithms, the Gymnasium-based API [@FaramaFoundation2023] should be used, which has been established as a standard in recent years.
+For other projects implementing the Gymnasium-based API, such as GridAlive (based on Grid2Op [@Donnot2020]) and ChroniX2Grid [@Marot2020], the focus remains on top-down control and steady-state models.
+Therefore, ``ElectricGrid.jl`` provides a tool to close these highlighted gaps.
 
 
 # Interfaces for control and reinforcement learning  
