### 2023-07-31 799ed56
Update paper.md
Minor corrections
