+++ README.md
@@ -24,23 +24,18 @@ ElectricGrid.jl is a library for setting up realistic electric grid simulations 
 ![ElectricGrid Framework](docs/src/assets/Overview_EG.png)
 
 ## Installation
-- Installation using the julia package manager (recommended):
+- Installation using the julia package manager (recommended if you want to use ElectricGrid in your project):
 In a julia terminal run the follwing:
 ```
-]
-activate .
-add ElectricGrid
+import Pkg
+Pkg.add("ElectricGrid")
 ```
 
-- Install from Github source:
+- Install from Github source (recommended if you want to run the example notebooks and scripts):
   - Clone the git and navigate to the directory
 ```
 git clone https://github.com/upb-lea/ElectricGrid.jl.git
 ```
-  - activate Julia
-
-  - activate the project by pressing `]`to access pkg mode and then `activate path/to/ElectricGrid` or `activate .` if you started julia in your ElectricGrid directory
-  - run `instantiate`
 
 ## Getting Started
 
