### 2023-08-01 790d48a
readme and docs update according to JOSS review
