+++ docs/src/index.md
@@ -22,7 +22,51 @@ Pkg.add("ElectricGrid")
 git clone https://github.com/upb-lea/ElectricGrid.jl.git
 ```
 
+## Statement of need
+
+Decentralised, electrical energy networks have special demands on operating and control procedures 
+to ensure a continuous and efficient energy supply and
+simultaneously play an important role in the integration of renewable energy sources ([@Guerrero2013](https://ieeexplore.ieee.org/document/6184305)). This applies both in connection with conventional centralised power grids and for islanded microgrids in remote areas ([@Lund2017](https://www.sciencedirect.com/science/article/abs/pii/S0360544217308812)).
+Due to their high efficiency and flexibility, power electronic converters have become the standard tool for integrating renewable energy sources, energy storages and loads in 
+electrical energy grids. 
+The field of power electronics covers the application of solid-state electronics to the control and
+conversion of electric power, which is performed with semiconductor switching 
+devices such as diodes or power transistors.
+This includes energy conversion in terms of voltage and current amplitude, frequency
+and phase angle, as well as the number of phases between two or more electrical energy systems to be connected.
+
+Controlling (decentralised) electric grids is a challenging task due to their stochastic, heterogeneous 
+and volatile characteristics (in particular regarding the connected loads).
+At the same time, high requirements are made with regard to aspects such as safety, quality and availability.
+This results in a high demand for comprehensive testing of new control concepts during their development phase and comparisons with the state
+of the art to ensure their feasibility.
+This applies in particular to emerging data-driven control approaches such as 
+reinforcement learning (RL), the stability and operating behavior of
+which cannot be evaluated a priori ([@Garcia2015](https://www.jmlr.org/papers/volume16/garcia15a/garcia15a.pdf)).
+Besides RL methods, being data-driven, result in a model-free and self-adaptive controller design with little human effort labeling them a promising tool for controlling unknown or changing systems targeting the above described challanges.
+However, there is a need for further research into the requirements for energy networks in terms of safety, robustness and availability before RL-based controllers can be used in real applications ([@ZhangRL2018](https://ieeexplore.ieee.org/abstract/document/8468674), [@GlavicRL2018](https://www.sciencedirect.com/science/article/pii/S2405896317317238)).
 
 
+``ElectricGrid.jl`` is a Julia package for setting up realistic electric grid simulations with support for control options. A number of parameters are made avaible to the user to evaluate the various control options. If no details are given, all parameters are generated automatically, either through randomness on a physically meaningful basis or by verified design methods.
+This enables both experts from the field of electrical energy networks to test certain configurations and 
+experts from the field of artificial intelligence to test new control approaches
+without any prior knowledge of electrical engineering.
+Therefore, ``ElectricGrid.jl`` is designed to be used by students, academics, and industrial researchers in the field of simulation and data-driven analysis of electrical energy systems. 
+The primary objective of the toolbox is to facilitate entry for new users into 
+the modeling, control, and testing of
+small to large scale electric power grids and to provide a platform on which different control methods (including RL) can be compared under defined conditions (benchmarks).
+
+The experiments are based on dynamic simulations in the time domain which allows for accurate control
+and test investigations during transients and steady state down to component level. 
+This is an essential difference to already available open-source solutions for the simulation of electrical energy grids like PyPSA ([@Brown2017](https://pypsa.readthedocs.io/en/latest/)), Powermodels.jl ([@Coffrin2018](https://arxiv.org/abs/1711.01728)) and pandapower 
[…]
