### 2023-08-01 #5616 @Webbah
Thank you once again @degleris1 and @kiranshila  for the very good reviews and hints!

We have adjusted the paper regarding your comments with introducing a "statement of need" section and with more content regarding RL.

Additionally we have extended the comparison to other software packages like recommended and adjusted the documentation hopefully to your satisfaction.

In fact, the choice of programming language was not trivial for us. After an extensive speed comparison of the basic software structure between Python and Julia, Julia emerged as the performance winner in terms of computation time.
Since it was supposed to be an open source project, we did not consider matlab.
The calculation time is very important, especially for the ad-hoc data generation that is required for RL training.
Therefore we finally decided to use Julia.

We are very grateful for the suggestions and hints on our Julia implementation (dict -> struct).
As this was our first Julia project, we are learning a lot and will consider the better performing ideas in the future project work.

Additionally we have updated the documentation and double checked that all parameters are documented (https://upb-lea.github.io/ElectricGrid.jl/dev/Default_Parameters/).

@osorensen, we hope that we have sufficiently addressed the points mentioned above.
