+++ examples/scripts/env_example_nonlinear.jl
@@ -1,5 +1,6 @@
 using ElectricGrid
 using PlotlyJS
+using BenchmarkTools
 
 CM = [0. 1.
      -1. 0.]
@@ -11,15 +12,14 @@ pf_load = 1
 v_rms = 230
 R_load, L_load, X, Z = ParallelLoadImpedance(S_load, pf_load, v_rms)
 
-# Which value is plottet
-value = 4
+value = 5
 function NonlinearInductance(drop, length)
-    # drop, is to how much it decreses with i to infty
+    # drop, is to how much it decreses with i to infinity
     # length, length is the point, where its half decresed
     return x->(1-(drop/100)*1/(1+exp(-(abs(x)-length)/2)))
 end
 
-l = NonlinearInductance(90,1)
+l = NonlinearInductance(95,1)
 
 parameters = Dict{Any, Any}(
     "source" => Any[
@@ -29,7 +29,7 @@ parameters = Dict{Any, Any}(
                     Dict{Any, Any}("impedance" => "R", "R" => R_load, "v_limit" => 1e4, "i_limit" => 1e4)
                     ],
     "cable"   => Any[
-                    Dict{Any, Any}("R" => 1e-3, "L" => x->1e-4*l(x), "C" => 1e-4),
+                    Dict{Any, Any}("R" => 1e-3, "L" => x->1e-4, "C" => 1e-4),
                     ],
     "grid" => Dict{Any, Any}("fs"=>1e4, "phase"=>3, "v_rms"=>230, "f_grid" => 50, "ramp_end"=>0.00)
 )
