+++ src/electric_grid_env.jl
@@ -19,7 +19,7 @@ mutable struct ElectricGridEnv <: AbstractEnv
     steps
     state_ids
     v_dc
-    v_dc_arr
+    vdc_link_voltages
     norm_array
     convert_state_to_cpu
     reward
@@ -274,57 +274,8 @@ function ElectricGridEnv(;
         end
     end
 
-    vdc_fixed = 0
-    v_dc = ones(nc.num_sources)  # vector to store evaluated v_dc_arr (constants and
-    # functions) in the env, needed e.g. in the DataHook
-    v_dc_arr = []  # array to store all functions for v_dc as well as constants
-    for (source_number, source) in enumerate(nc.parameters["source"])
-        if haskey(source, "source_type")
-            if source["source_type"] == "ideal"
-                # set vdc for that source
-                if haskey(source, "vdc")
-                    v_dc[source_number] = source["vdc"]
-                    fun = (env, G, T) -> source["vdc"]
-                    push!(v_dc_arr, fun)
-                else
-                    v_dc[source_number] = 800
-                    fun = (env, G, T) -> 800
-                    push!(v_dc_arr, fun)
-                    vdc_fixed += 1
-                end
-            elseif source["source_type"] == "pv"
-                #v_dc[source_number] = :(GetV(SolarArray, env.x[1]*env.action, G, T))
-                #TODO : how to calculate i_dc in 3-phase grid? Which current of env to use?
-                #TODO : $source_number does only fit if all L filters! Otherwise how to
-                #       define the offet for $source_number?!?!?
-                # TODO built pv module from parameter dict - where to define? In env?
-                ModulePV = SolarModule()
-                SolarArr = SolarArray(; ModuleParameters=ModulePV)
-                # find(x -> .... source$source_number_i_L in state_ids)
-                fun = (env, G, T) -> GetV(:($SolarArr),
-                    env.x[:($source_number)] * env.action, G, T)
-                push!(v_dc_arr, fun)
-
-                # first value set to 0
-                v_dc[source_number] = 0
-            else
-                @warn "sourceType not known! vdc set to fixed value"
-                v_dc[source_number] = 800
-                fun = (env, G, T) -> 800
-                push!(v_dc_arr, fun)
-                vdc_fixed += 1
-            end
-        else
-            @warn "sourceType not defined! vdc set to fixed value, if not wanted please
-                   define nc.parameters -> source -> source_type (e.g. = ideal)"
-            v_dc[source_number] = 800
-            fun = (env, G, T) -> 800
-            push!(v_dc_arr, fun)
-            vdc_fixed += 1
-        end
-    end
-    vdc_fixed > 0 && @warn("$vdc_fixed DC-link voltages set to 800 V - please define in
-                            nc.parameters -> source -> vdc.")
+    vdc_link_voltages =  DCLinkVoltagesInit(nc) # create dc link voltage vector
+    v_dc = zeros(num_sources) # get initial v_dc
 
     if verbosity > 1
         @info "Normalization is done based on the defined parameter limits."
@@ -418,12 +369,44 @@ function ElectricGridEnv(;
 
     y = (A * Vector(x) + B * (Vector(action))) .* (state_parameters)
 
-    ElectricGridEnv(verbosity, nc, sys_d, action_space, state_space,
-        false, inner_featurize, featurize, prepare_action, reward_function,
-        x0, x, t0, t, ts, state, maxsteps, 0, state_ids,
-        v_dc, v_dc_arr, norm_array, convert_state_to_cpu,
[…]
