### 2023-08-16 4092819
Include variable_dc_link.jl
