+++ src/ElectricGrid.jl
@@ -33,6 +33,7 @@ include("./node_constructor.jl")
 include("./custom_control.jl")
 include("./solar_module.jl")
 include("./battery.jl")
+include("./varaible_dc_link.jl")
 include("./electric_grid_env.jl")
 include("./agent_ddpg.jl")
 include("./classical_control.jl")
