### 2023-08-17 79b2be8
Info for the solver
