### 2023-08-18 da8067b
Deleted Prints and added stuff to nb
