### 2023-08-27 #5616 @osorensen
> ```
> Reference check summary (note 'MISSING' DOIs are suggestions that need verification):
>
> OK DOIs
>
> - 10.23919/PSCC.2018.8442948 is OK
> - 10.1109/TIE.2012.2194969 is OK
> - 10.1016/j.energy.2017.05.123 is OK
> - 10.17775/CSEEJPES.2018.00520 is OK
>
> MISSING DOIs
>
> - 10.1016/j.ifacol.2017.08.1217 may be a valid DOI for title: Reinforcement Learning for Electric Power System Decision and Control: Past Considerations and Perspectives
> - 10.5334/jors.188 may be a valid DOI for title: PyPSA: Python for power system analysis
> - 10.1109/tpwrs.2018.2829021 may be a valid DOI for title: pandapower—an open-source python tool for convenient modeling, analysis, and optimization of electric power systems
> - 10.1109/tpwrs.2023.3303291 may be a valid DOI for title: Revisiting Power Systems Time-domain Simulation Methods and Models
>
> INVALID DOIs
>
> - None
> ```

@Webbah, could you please check if the missing DOIs suggested here are correct, and if so, add them.
