### 2023-08-27 upb-lea/ElectricGrid.jl#112 @osorensen
Editorial comments on JOSS submission
Whenever there are multiple references, please put them into the same set of parentheses using `[@key1;@key2]`, rather than citing each separately. This comment applies to lines 40-41, cf. screenshot below:

https://github.com/openjournals/joss-reviews/issues/5616
