### 2023-08-27 upb-lea/ElectricGrid.jl#113 @osorensen
Editorial comment on JOSS submission
The sentence highlighted below does not seem complete. What do you mean by "tools like (Lara et al., 2023)"?

https://github.com/openjournals/joss-reviews/issues/5616
