### 2023-08-27 upb-lea/ElectricGrid.jl#113 @osorensen
If you mean "tools like the one presented in Lara et al. (2023)", then please rewrite to that. If you need it, note that you can create references without paranetheses with `@key`, i.e., dropping the square brackets.
