### 2023-08-27 upb-lea/ElectricGrid.jl#114 @osorensen
Editorial comments on JOSS submission
Please refer to the figures (e.g., "See Figure 2"), rather than stating "the figure below". This applies to both Figure 1 and Figure 2.

https://github.com/openjournals/joss-reviews/issues/5616
