### 2023-08-29 24b459c
Editorial comments on JOSS submission #112 changed multiple references
