### 2023-08-29 upb-lea/ElectricGrid.jl#112 @Webbah
Thank you for the hint.
I have changed the mutliple references and cross checked for others
