### 2023-08-29 1f7bf50
Editorial comment on JOSS submission #113 Edit sentence and reference
