### 2023-08-29 upb-lea/ElectricGrid.jl#113 @Webbah
Yes, that was our intention and I am with you that the sentence is a bit "short".
Thank you for the suggestions, changes have been updated
