### 2023-08-29 a4958c1
Editorial comments on JOSS submission #114 changed figure reference
