### 2023-08-29 upb-lea/ElectricGrid.jl#114 @Webbah
Thank you for the suggestion. Changes have been made accordingly
