### 2023-08-29 7b4b8fb
JOSS paper #44 added DOIs
