+++ docs/JOSS/paper.bib
@@ -73,7 +73,8 @@
   volume = {50},
   year = {2017},
   author = {Mevludin Glavic and Raphaël Fonteneau and Damien Ernst},
-  keywords = {Electric power system, reinforcement learning, control, decision}
+  keywords = {Electric power system, reinforcement learning, control, decision},
+  doi={10.1016/j.ifacol.2017.08.1217}
   }
 
 @Article{Brown2017,
@@ -81,6 +82,7 @@
   journal = {arXiv preprint arXiv:1707.09913},
   title   = {PyPSA: Python for power system analysis},
   year    = {2017},
+  doi={https://doi.org/10.5334/jors.188}
 }
 
 @Article{Thurner2018,
@@ -92,13 +94,15 @@
   pages     = {6510--6521},
   volume    = {33},
   publisher = {IEEE},
+  doi={10.1109/TPWRS.2018.2829021}
 }
 
    @article{lara2023revisiting,
     title={Revisiting Power Systems Time-domain Simulation Methods and Models},
     author={Lara, Jose Daniel and Henriquez-Auba, Rodrigo and Ramasubramanian, Deepak and Dhople, Sairaj and Callaway, Duncan S and Sanders, Seth},
     journal={arXiv preprint arXiv:2301.10043},
-    year={2023}
+    year={2023},
+    doi={https://doi.org/10.1109/TPWRS.2023.3303291}
 }
 
 @Misc{FaramaFoundation2023,
