+++ docs/JOSS/paper.bib
@@ -82,7 +82,7 @@
   journal = {arXiv preprint arXiv:1707.09913},
   title   = {PyPSA: Python for power system analysis},
   year    = {2017},
-  doi={https://doi.org/10.5334/jors.188}
+  doi={10.5334/jors.188}
 }
 
 @Article{Thurner2018,
@@ -102,7 +102,7 @@
     author={Lara, Jose Daniel and Henriquez-Auba, Rodrigo and Ramasubramanian, Deepak and Dhople, Sairaj and Callaway, Duncan S and Sanders, Seth},
     journal={arXiv preprint arXiv:2301.10043},
     year={2023},
-    doi={https://doi.org/10.1109/TPWRS.2023.3303291}
+    doi={10.1109/TPWRS.2023.3303291}
 }
 
 @Misc{FaramaFoundation2023,
