### 2023-08-29 #5616 @Webbah
- v1.0.2 is the release of the software with the latest changes from the review
- DOI: 10.5281/zenodo.8297533](https://doi.org/10.5281/zenodo.8297533
