### 2023-08-29 #5616 @osorensen
Here's an example of a recently published JOSS paper: https://zenodo.org/record/8280775
