### 2023-09-01 7965dc3
Feature 100 additional rl (#110)
* initital file to start working with

* bugfixes

* update

* #100 add td3 agent

* #100 add test scripts

* #100 convert action space using wrappers

* #100 add experiments with td3

* fix `TD3Policy` not defined

* adapt td3 to multi dims action and state

* #100 log first results with TD3

* #100 add wandb support

* #100 add notebook example for `TD3`

* #100 fix `max_steps` not defined

* #100 re organise

* fix spelling

* #100 add wandb logged plots

* add wandb, statistcs, revise to project toml

* #100 rm wandb

* #100 fix multiple methods with same n
[…]
